@@ -2,938 +2,13 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net8.0</TargetFramework>
-    <Version>1.2.0</Version>
+    <TargetFramework>net9.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
-    <None Remove="fonts\1Row.flf" />
-    <None Remove="fonts\3-D.flf" />
-    <None Remove="fonts\3D Diagonal.flf" />
-    <None Remove="fonts\3D-ASCII.flf" />
-    <None Remove="fonts\3d.flf" />
-    <None Remove="fonts\3d_diagonal.flf" />
-    <None Remove="fonts\3x5.flf" />
-    <None Remove="fonts\4Max.flf" />
-    <None Remove="fonts\5 Line Oblique.flf" />
-    <None Remove="fonts\5lineoblique.flf" />
-    <None Remove="fonts\646-ca.flc" />
-    <None Remove="fonts\646-ca2.flc" />
-    <None Remove="fonts\646-cn.flc" />
-    <None Remove="fonts\646-cu.flc" />
-    <None Remove="fonts\646-de.flc" />
-    <None Remove="fonts\646-dk.flc" />
-    <None Remove="fonts\646-es.flc" />
-    <None Remove="fonts\646-es2.flc" />
-    <None Remove="fonts\646-fr.flc" />
-    <None Remove="fonts\646-gb.flc" />
-    <None Remove="fonts\646-hu.flc" />
-    <None Remove="fonts\646-irv.flc" />
-    <None Remove="fonts\646-it.flc" />
-    <None Remove="fonts\646-jp.flc" />
-    <None Remove="fonts\646-kr.flc" />
-    <None Remove="fonts\646-no.flc" />
-    <None Remove="fonts\646-no2.flc" />
-    <None Remove="fonts\646-pt.flc" />
-    <None Remove="fonts\646-pt2.flc" />
-    <None Remove="fonts\646-se.flc" />
-    <None Remove="fonts\646-se2.flc" />
-    <None Remove="fonts\646-yu.flc" />
-    <None Remove="fonts\8859-2.flc" />
-    <None Remove="fonts\8859-3.flc" />
-    <None Remove="fonts\8859-4.flc" />
-    <None Remove="fonts\8859-5.flc" />
-    <None Remove="fonts\8859-7.flc" />
-    <None Remove="fonts\8859-8.flc" />
-    <None Remove="fonts\8859-9.flc" />
-    <None Remove="fonts\Acrobatic.flf" />
-    <None Remove="fonts\Alligator.flf" />
-    <None Remove="fonts\Alligator2.flf" />
-    <None Remove="fonts\alligator3.flf" />
-    <None Remove="fonts\Alpha.flf" />
-    <None Remove="fonts\Alphabet.flf" />
-    <None Remove="fonts\AMC 3 Line.flf" />
-    <None Remove="fonts\AMC 3 Liv1.flf" />
-    <None Remove="fonts\AMC AAA01.flf" />
-    <None Remove="fonts\AMC Neko.flf" />
-    <None Remove="fonts\AMC Razor.flf" />
-    <None Remove="fonts\AMC Razor2.flf" />
-    <None Remove="fonts\AMC Slash.flf" />
-    <None Remove="fonts\AMC Slider.flf" />
-    <None Remove="fonts\AMC Thin.flf" />
-    <None Remove="fonts\AMC Tubes.flf" />
-    <None Remove="fonts\AMC Untitled.flf" />
-    <None Remove="fonts\amc3line.flf" />
-    <None Remove="fonts\amc3liv1.flf" />
-    <None Remove="fonts\amcaaa01.flf" />
-    <None Remove="fonts\amcneko.flf" />
-    <None Remove="fonts\amcrazo2.flf" />
-    <None Remove="fonts\amcrazor.flf" />
-    <None Remove="fonts\amcslash.flf" />
-    <None Remove="fonts\amcslder.flf" />
-    <None Remove="fonts\amcthin.flf" />
-    <None Remove="fonts\amctubes.flf" />
-    <None Remove="fonts\amcun1.flf" />
-    <None Remove="fonts\ANSI Regular.flf" />
-    <None Remove="fonts\ANSI Shadow.flf" />
-    <None Remove="fonts\Arrows.flf" />
-    <None Remove="fonts\ASCII New Roman.flf" />
-    <None Remove="fonts\ascii12.tlf" />
-    <None Remove="fonts\ascii9.tlf" />
-    <None Remove="fonts\ascii_new_roman.flf" />
-    <None Remove="fonts\Avatar.flf" />
-    <None Remove="fonts\B1FF.flf" />
-    <None Remove="fonts\B1FF.php" />
-    <None Remove="fonts\Banner.flf" />
-    <None Remove="fonts\Banner3-D.flf" />
-    <None Remove="fonts\Banner3.flf" />
-    <None Remove="fonts\Banner4.flf" />
-    <None Remove="fonts\Barbwire.flf" />
-    <None Remove="fonts\Basic.flf" />
-    <None Remove="fonts\Bear.flf" />
-    <None Remove="fonts\Bell.flf" />
-    <None Remove="fonts\Benjamin.flf" />
-    <None Remove="fonts\Big Chief.flf" />
-    <None Remove="fonts\Big Money-ne.flf" />
-    <None Remove="fonts\Big Money-nw.flf" />
-    <None Remove="fonts\Big Money-se.flf" />
-    <None Remove="fonts\Big Money-sw.flf" />
-    <None Remove="fonts\Big.flf" />
-    <None Remove="fonts\bigascii12.tlf" />
-    <None Remove="fonts\bigascii9.tlf" />
-    <None Remove="fonts\bigchief.flf" />
-    <None Remove="fonts\Bigfig.flf" />
-    <None Remove="fonts\bigmono12.tlf" />
-    <None Remove="fonts\bigmono9.tlf" />
-    <None Remove="fonts\Binary.flf" />
-    <None Remove="fonts\Block.flf" />
-    <None Remove="fonts\Blocks.flf" />
-    <None Remove="fonts\Bloody.flf" />
-    <None Remove="fonts\Bolger.flf" />
-    <None Remove="fonts\Braced.flf" />
-    <None Remove="fonts\Bright.flf" />
-    <None Remove="fonts\Broadway KB.flf" />
-    <None Remove="fonts\Broadway.flf" />
-    <None Remove="fonts\broadway_kb.flf" />
-    <None Remove="fonts\Bubble.flf" />
-    <None Remove="fonts\Bulbhead.flf" />
-    <None Remove="fonts\calgphy2.flf" />
-    <None Remove="fonts\Caligraphy.flf" />
-    <None Remove="fonts\Caligraphy2.flf" />
-    <None Remove="fonts\Calvin S.flf" />
-    <None Remove="fonts\Cards.flf" />
-    <None Remove="fonts\Catwalk.flf" />
-    <None Remove="fonts\Chiseled.flf" />
-    <None Remove="fonts\Chunky.flf" />
-    <None Remove="fonts\circle.tlf" />
-    <None Remove="fonts\Coinstak.flf" />
-    <None Remove="fonts\Cola.flf" />
-    <None Remove="fonts\Colossal.flf" />
-    <None Remove="fonts\Computer.flf" />
-    <None Remove="fonts\Contessa.flf" />
-    <None Remove="fonts\Contrast.flf" />
-    <None Remove="fonts\cosmic.flf" />
-    <None Remove="fonts\Cosmike.flf" />
-    <None Remove="fonts\Crawford.flf" />
-    <None Remove="fonts\Crawford2.flf" />
-    <None Remove="fonts\Crazy.flf" />
-    <None Remove="fonts\Cricket.flf" />
-    <None Remove="fonts\Cursive.flf" />
-    <None Remove="fonts\Cyberlarge.flf" />
-    <None Remove="fonts\Cybermedium.flf" />
-    <None Remove="fonts\Cybersmall.flf" />
-    <None Remove="fonts\Cygnet.flf" />
-    <None Remove="fonts\DANC4.flf" />
-    <None Remove="fonts\Dancing Font.flf" />
-    <None Remove="fonts\dancingfont.flf" />
-    <None Remove="fonts\Decimal.flf" />
-    <None Remove="fonts\Def Leppard.flf" />
-    <None Remove="fonts\defleppard.flf" />
-    <None Remove="fonts\Delta Corps Priest 1.flf" />
-    <None Remove="fonts\Diamond.flf" />
-    <None Remove="fonts\Diet Cola.flf" />
-    <None Remove="fonts\dietcola.flf" />
-    <None Remove="fonts\Digital.flf" />
-    <None Remove="fonts\Doh.flf" />
-    <None Remove="fonts\Doom.flf" />
-    <None Remove="fonts\DOS Rebel.flf" />
-    <None Remove="fonts\dosrebel.flf" />
-    <None Remove="fonts\Dot Matrix.flf" />
-    <None Remove="fonts\dotmatrix.flf" />
-    <None Remove="fonts\Double Shorts.flf" />
-    <None Remove="fonts\Double.flf" />
-    <None Remove="fonts\doubleshorts.flf" />
-    <None Remove="fonts\Dr Pepper.flf" />
-    <None Remove="fonts\drpepper.flf" />
-    <None Remove="fonts\DWhistled.flf" />
-    <None Remove="fonts\Efti Chess.flf" />
-    <None Remove="fonts\Efti Font.flf" />
-    <None Remove="fonts\Efti Italic.flf" />
-    <None Remove="fonts\Efti Piti.flf" />
-    <None Remove="fonts\Efti Robot.flf" />
-    <None Remove="fonts\Efti Wall.flf" />
-    <None Remove="fonts\Efti Water.flf" />
-    <None Remove="fonts\eftichess.flf" />
-    <None Remove="fonts\eftifont.flf" />
-    <None Remove="fonts\eftipiti.flf" />
-    <None Remove="fonts\eftirobot.flf" />
-    <None Remove="fonts\eftitalic.flf" />
-    <None Remove="fonts\eftiwall.flf" />
-    <None Remove="fonts\eftiwater.flf" />
-    <None Remove="fonts\Electronic.flf" />
-    <None Remove="fonts\Elite.flf" />
-    <None Remove="fonts\emboss.tlf" />
-    <None Remove="fonts\emboss2.tlf" />
-    <None Remove="fonts\Epic.flf" />
-    <None Remove="fonts\Fender.flf" />
-    <None Remove="fonts\Filter.flf" />
-    <None Remove="fonts\Fire Font-k.flf" />
-    <None Remove="fonts\Fire Font-s.flf" />
-    <None Remove="fonts\fire_font-k.flf" />
-    <None Remove="fonts\fire_font-s.flf" />
-    <None Remove="fonts\Flipped.flf" />
-    <None Remove="fonts\Flower Power.flf" />
-    <None Remove="fonts\flowerpower.flf" />
-    <None Remove="fonts\Four Tops.flf" />
-    <None Remove="fonts\fourtops.flf" />
-    <None Remove="fonts\Fraktur.flf" />
-    <None Remove="fonts\frango.flc" />
-    <None Remove="fonts\Fun Face.flf" />
-    <None Remove="fonts\Fun Faces.flf" />
-    <None Remove="fonts\funface.flf" />
-    <None Remove="fonts\funfaces.flf" />
-    <None Remove="fonts\future.tlf" />
-    <None Remove="fonts\Fuzzy.flf" />
-    <None Remove="fonts\Georgi16.flf" />
-    <None Remove="fonts\Georgia11.flf" />
-    <None Remove="fonts\Ghost.flf" />
-    <None Remove="fonts\Ghoulish.flf" />
-    <None Remove="fonts\Glenyn.flf" />
-    <None Remove="fonts\Goofy.flf" />
-    <None Remove="fonts\Gothic.flf" />
-    <None Remove="fonts\Graceful.flf" />
-    <None Remove="fonts\Gradient.flf" />
-    <None Remove="fonts\Graffiti.flf" />
-    <None Remove="fonts\Greek.flf" />
-    <None Remove="fonts\Heart Left.flf" />
-    <None Remove="fonts\Heart Right.flf" />
-    <None Remove="fonts\heart_left.flf" />
-    <None Remove="fonts\heart_right.flf" />
-    <None Remove="fonts\Henry 3D.flf" />
-    <None Remove="fonts\henry3d.flf" />
-    <None Remove="fonts\Hex.flf" />
-    <None Remove="fonts\Hieroglyphs.flf" />
-    <None Remove="fonts\Hollywood.flf" />
-    <None Remove="fonts\Horizontal Left.flf" />
-    <None Remove="fonts\Horizontal Right.flf" />
-    <None Remove="fonts\horizontalleft.flf" />
-    <None Remove="fonts\horizontalright.flf" />
-    <None Remove="fonts\hz.flc" />
-    <None Remove="fonts\ICL-1900.flf" />
-    <None Remove="fonts\ilhebrew.flc" />
-    <None Remove="fonts\Impossible.flf" />
-    <None Remove="fonts\Invita.flf" />
-    <None Remove="fonts\Isometric1.flf" />
-    <None Remove="fonts\Isometric2.flf" />
-    <None Remove="fonts\Isometric3.flf" />
-    <None Remove="fonts\Isometric4.flf" />
-    <None Remove="fonts\Italic.flf" />
-    <None Remove="fonts\Ivrit.flf" />
-    <None Remove="fonts\Jacky.flf" />
-    <None Remove="fonts\Jazmine.flf" />
-    <None Remove="fonts\Jerusalem.flf" />
-    <None Remove="fonts\jis0201.flc" />
-    <None Remove="fonts\JS Block Letters.flf" />
-    <None Remove="fonts\JS Bracket Letters.flf" />
-    <None Remove="fonts\JS Capital Curves.flf" />
-    <None Remove="fonts\JS Cursive.flf" />
-    <None Remove="fonts\JS Stick Letters.flf" />
-    <None Remove="fonts\Katakana.flf" />
-    <None Remove="fonts\Kban.flf" />
-    <None Remove="fonts\Keyboard.flf" />
-    <None Remove="fonts\Knob.flf" />
-    <None Remove="fonts\koi8r.flc" />
-    <None Remove="fonts\Konto Slant.flf" />
-    <None Remove="fonts\Konto.flf" />
-    <None Remove="fonts\kontoslant.flf" />
-    <None Remove="fonts\Larry 3D 2.flf" />
-    <None Remove="fonts\Larry 3D.flf" />
-    <None Remove="fonts\larry3d.flf" />
-    <None Remove="fonts\LCD.flf" />
-    <None Remove="fonts\Lean.flf" />
-    <None Remove="fonts\letter.tlf" />
-    <None Remove="fonts\Letters.flf" />
-    <None Remove="fonts\Lil Devil.flf" />
-    <None Remove="fonts\lildevil.flf" />
-    <None Remove="fonts\Line Blocks.flf" />
-    <None Remove="fonts\lineblocks.flf" />
-    <None Remove="fonts\Linux.flf" />
-    <None Remove="fonts\Lockergnome.flf" />
-    <None Remove="fonts\Madrid.flf" />
-    <None Remove="fonts\Marquee.flf" />
-    <None Remove="fonts\Maxfour.flf" />
-    <None Remove="fonts\maxiwi.flf" />
-    <None Remove="fonts\Merlin1.flf" />
-    <None Remove="fonts\Merlin2.flf" />
-    <None Remove="fonts\Mike.flf" />
-    <None Remove="fonts\Mini.flf" />
-    <None Remove="fonts\miniwi.flf" />
-    <None Remove="fonts\Mirror.flf" />
-    <None Remove="fonts\Mnemonic.flf" />
-    <None Remove="fonts\Modular.flf" />
-    <None Remove="fonts\mono12.tlf" />
-    <None Remove="fonts\mono9.tlf" />
-    <None Remove="fonts\Morse.flf" />
-    <None Remove="fonts\Morse2.flf" />
-    <None Remove="fonts\moscow.flc" />
-    <None Remove="fonts\Moscow.flf" />
-    <None Remove="fonts\Mshebrew210.flf" />
-    <None Remove="fonts\Muzzle.flf" />
-    <None Remove="fonts\Nancyj-Fancy.flf" />
-    <None Remove="fonts\Nancyj-Improved.flf" />
-    <None Remove="fonts\Nancyj-Underlined.flf" />
-    <None Remove="fonts\Nancyj.flf" />
-    <None Remove="fonts\Nipples.flf" />
-    <None Remove="fonts\NScript.flf" />
-    <None Remove="fonts\NT Greek.flf" />
-    <None Remove="fonts\ntgreek.flf" />
-    <None Remove="fonts\NV Script.flf" />
-    <None Remove="fonts\O8.flf" />
-    <None Remove="fonts\Octal.flf" />
-    <None Remove="fonts\Ogre.flf" />
-    <None Remove="fonts\Old Banner.flf" />
-    <None Remove="fonts\oldbanner.flf" />
-    <None Remove="fonts\OS2.flf" />
-    <None Remove="fonts\pagga.tlf" />
-    <None Remove="fonts\Patorjk%27s Cheese.flf" />
-    <None Remove="fonts\Patorjk-HeX.flf" />
-    <None Remove="fonts\Pawp.flf" />
-    <None Remove="fonts\Peaks Slant.flf" />
-    <None Remove="fonts\Peaks.flf" />
-    <None Remove="fonts\peaksslant.flf" />
-    <None Remove="fonts\Pebbles.flf" />
-    <None Remove="fonts\Pepper.flf" />
-    <None Remove="fonts\Poison.flf" />
-    <None Remove="fonts\Puffy.flf" />
-    <None Remove="fonts\Puzzle.flf" />
-    <None Remove="fonts\Pyramid.flf" />
-    <None Remove="fonts\Rammstein.flf" />
-    <None Remove="fonts\README.md" />
-    <None Remove="fonts\rebel.tlf" />
-    <None Remove="fonts\Rectangles.flf" />
-    <None Remove="fonts\Red Phoenix.flf" />
-    <None Remove="fonts\red_phoenix.flf" />
-    <None Remove="fonts\Relief.flf" />
-    <None Remove="fonts\Relief2.flf" />
-    <None Remove="fonts\rev.flf" />
-    <None Remove="fonts\Reverse.flf" />
-    <None Remove="fonts\Roman.flf" />
-    <None Remove="fonts\Rot13.flf" />
-    <None Remove="fonts\Rot13.php" />
-    <None Remove="fonts\Rotated.flf" />
-    <None Remove="fonts\Rounded.flf" />
-    <None Remove="fonts\Rowan Cap.flf" />
-    <None Remove="fonts\rowancap.flf" />
-    <None Remove="fonts\Rozzo.flf" />
-    <None Remove="fonts\Runic.flf" />
-    <None Remove="fonts\Runyc.flf" />
-    <None Remove="fonts\rusto.tlf" />
-    <None Remove="fonts\rustofat.tlf" />
-    <None Remove="fonts\S Blood.flf" />
-    <None Remove="fonts\s-relief.flf" />
-    <None Remove="fonts\Santa Clara.flf" />
-    <None Remove="fonts\santaclara.flf" />
-    <None Remove="fonts\sblood.flf" />
-    <None Remove="fonts\Script.flf" />
-    <None Remove="fonts\Serifcap.flf" />
-    <None Remove="fonts\Shadow.flf" />
-    <None Remove="fonts\Shimrod.flf" />
-    <None Remove="fonts\Short.flf" />
-    <None Remove="fonts\SL Script.flf" />
-    <None Remove="fonts\Slant Relief.flf" />
-    <None Remove="fonts\Slant.flf" />
-    <None Remove="fonts\Slide.flf" />
-    <None Remove="fonts\slscript.flf" />
-    <None Remove="fonts\Small Caps.flf" />
-    <None Remove="fonts\Small Isometric1.flf" />
-    <None Remove="fonts\Small Keyboard.flf" />
-    <None Remove="fonts\Small Poison.flf" />
-    <None Remove="fonts\Small Script.flf" />
-    <None Remove="fonts\Small Shadow.flf" />
-    <None Remove="fonts\Small Slant.flf" />
-    <None Remove="fonts\Small Tengwar.flf" />
-    <None Remove="fonts\Small.flf" />
-    <None Remove="fonts\smallcaps.flf" />
-    <None Remove="fonts\smascii12.tlf" />
-    <None Remove="fonts\smascii9.tlf" />
-    <None Remove="fonts\smblock.tlf" />
-    <None Remove="fonts\smbraille.tlf" />
-    <None Remove="fonts\smisome1.flf" />
-    <None Remove="fonts\smkeyboard.flf" />
-    <None Remove="fonts\smmono12.tlf" />
-    <None Remove="fonts\smmono9.tlf" />
-    <None Remove="fonts\smpoison.flf" />
-    <None Remove="fonts\smscript.flf" />
-    <None Remove="fonts\smshadow.flf" />
-    <None Remove="fonts\smslant.flf" />
-    <None Remove="fonts\smtengwar.flf" />
-    <None Remove="fonts\Soft.flf" />
-    <None Remove="fonts\Speed.flf" />
-    <None Remove="fonts\Spliff.flf" />
-    <None Remove="fonts\Stacey.flf" />
-    <None Remove="fonts\Stampate.flf" />
-    <None Remove="fonts\Stampatello.flf" />
-    <None Remove="fonts\Standard.flf" />
-    <None Remove="fonts\Star Strips.flf" />
-    <None Remove="fonts\Star Wars.flf" />
-    <None Remove="fonts\starstrips.flf" />
-    <None Remove="fonts\starwars.flf" />
-    <None Remove="fonts\Stellar.flf" />
-    <None Remove="fonts\Stforek.flf" />
-    <None Remove="fonts\Stick Letters.flf" />
-    <None Remove="fonts\Stop.flf" />
-    <None Remove="fonts\Straight.flf" />
-    <None Remove="fonts\Stronger Than All.flf" />
-    <None Remove="fonts\Sub-Zero.flf" />
-    <None Remove="fonts\Swamp Land.flf" />
-    <None Remove="fonts\swampland.flf" />
-    <None Remove="fonts\Swan.flf" />
-    <None Remove="fonts\Sweet.flf" />
-    <None Remove="fonts\Tanja.flf" />
-    <None Remove="fonts\Tengwar.flf" />
-    <None Remove="fonts\Term.flf" />
-    <None Remove="fonts\Test1.flf" />
-    <None Remove="fonts\The Edge.flf" />
-    <None Remove="fonts\Thick.flf" />
-    <None Remove="fonts\Thin.flf" />
-    <None Remove="fonts\THIS.flf" />
-    <None Remove="fonts\Thorned.flf" />
-    <None Remove="fonts\Three Point.flf" />
-    <None Remove="fonts\threepoint.flf" />
-    <None Remove="fonts\Ticks Slant.flf" />
-    <None Remove="fonts\Ticks.flf" />
-    <None Remove="fonts\ticksslant.flf" />
-    <None Remove="fonts\Tiles.flf" />
-    <None Remove="fonts\Tinker-Toy.flf" />
-    <None Remove="fonts\Tombstone.flf" />
-    <None Remove="fonts\Train.flf" />
-    <None Remove="fonts\Trek.flf" />
-    <None Remove="fonts\Tsalagi.flf" />
-    <None Remove="fonts\Tubular.flf" />
-    <None Remove="fonts\Twisted.flf" />
-    <None Remove="fonts\Two Point.flf" />
-    <None Remove="fonts\twopoint.flf" />
-    <None Remove="fonts\Univers.flf" />
-    <None Remove="fonts\upper.flc" />
-    <None Remove="fonts\USA Flag.flf" />
-    <None Remove="fonts\usaflag.flf" />
-    <None Remove="fonts\ushebrew.flc" />
-    <None Remove="fonts\uskata.flc" />
-    <None Remove="fonts\utf8.flc" />
-    <None Remove="fonts\Varsity.flf" />
-    <None Remove="fonts\Wavy.flf" />
-    <None Remove="fonts\Weird.flf" />
-    <None Remove="fonts\Wet Letter.flf" />
-    <None Remove="fonts\wetletter.flf" />
-    <None Remove="fonts\Whimsy.flf" />
-    <None Remove="fonts\wideterm.tlf" />
-    <None Remove="fonts\Wow.flf" />
-    <None Remove="images\FinalLogo.jpg" />
-    <None Remove="images\star1.png" />
-    <None Remove="images\star2.jpg" />
-    <None Remove="images\star3.png" />
-    <None Remove="images\star4.png" />
-    <None Remove="images\star5.png" />
-    <None Remove="images\star6.jpg" />
-    <None Remove="images\star6b.jpg" />
-    <None Remove="Lib\Spectre.Console.deps.json" />
-    <None Remove="Lib\Spectre.Console.xml" />
-    <None Remove="Spectre.Console.deps.json" />
-    <None Remove="Spectre.Console.xml" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Content Include="fonts\1Row.flf" />
-    <Content Include="fonts\3-D.flf" />
-    <Content Include="fonts\3D Diagonal.flf" />
-    <Content Include="fonts\3D-ASCII.flf" />
-    <Content Include="fonts\3d.flf" />
-    <Content Include="fonts\3d_diagonal.flf" />
-    <Content Include="fonts\3x5.flf" />
-    <Content Include="fonts\4Max.flf" />
-    <Content Include="fonts\5 Line Oblique.flf" />
-    <Content Include="fonts\5lineoblique.flf" />
-    <Content Include="fonts\646-ca.flc" />
-    <Content Include="fonts\646-ca2.flc" />
-    <Content Include="fonts\646-cn.flc" />
-    <Content Include="fonts\646-cu.flc" />
-    <Content Include="fonts\646-de.flc" />
-    <Content Include="fonts\646-dk.flc" />
-    <Content Include="fonts\646-es.flc" />
-    <Content Include="fonts\646-es2.flc" />
-    <Content Include="fonts\646-fr.flc" />
-    <Content Include="fonts\646-gb.flc" />
-    <Content Include="fonts\646-hu.flc" />
-    <Content Include="fonts\646-irv.flc" />
-    <Content Include="fonts\646-it.flc" />
-    <Content Include="fonts\646-jp.flc" />
-    <Content Include="fonts\646-kr.flc" />
-    <Content Include="fonts\646-no.flc" />
-    <Content Include="fonts\646-no2.flc" />
-    <Content Include="fonts\646-pt.flc" />
-    <Content Include="fonts\646-pt2.flc" />
-    <Content Include="fonts\646-se.flc" />
-    <Content Include="fonts\646-se2.flc" />
-    <Content Include="fonts\646-yu.flc" />
-    <Content Include="fonts\8859-2.flc" />
-    <Content Include="fonts\8859-3.flc" />
-    <Content Include="fonts\8859-4.flc" />
-    <Content Include="fonts\8859-5.flc" />
-    <Content Include="fonts\8859-7.flc" />
-    <Content Include="fonts\8859-8.flc" />
-    <Content Include="fonts\8859-9.flc" />
-    <Content Include="fonts\Acrobatic.flf" />
-    <Content Include="fonts\Alligator.flf" />
-    <Content Include="fonts\Alligator2.flf" />
-    <Content Include="fonts\alligator3.flf" />
-    <Content Include="fonts\Alpha.flf" />
-    <Content Include="fonts\Alphabet.flf" />
-    <Content Include="fonts\AMC 3 Line.flf" />
-    <Content Include="fonts\AMC 3 Liv1.flf" />
-    <Content Include="fonts\AMC AAA01.flf" />
-    <Content Include="fonts\AMC Neko.flf" />
-    <Content Include="fonts\AMC Razor.flf" />
-    <Content Include="fonts\AMC Razor2.flf" />
-    <Content Include="fonts\AMC Slash.flf" />
-    <Content Include="fonts\AMC Slider.flf" />
-    <Content Include="fonts\AMC Thin.flf" />
-    <Content Include="fonts\AMC Tubes.flf" />
-    <Content Include="fonts\AMC Untitled.flf" />
-    <Content Include="fonts\amc3line.flf" />
-    <Content Include="fonts\amc3liv1.flf" />
-    <Content Include="fonts\amcaaa01.flf" />
-    <Content Include="fonts\amcneko.flf" />
-    <Content Include="fonts\amcrazo2.flf" />
-    <Content Include="fonts\amcrazor.flf" />
-    <Content Include="fonts\amcslash.flf" />
-    <Content Include="fonts\amcslder.flf" />
-    <Content Include="fonts\amcthin.flf" />
-    <Content Include="fonts\amctubes.flf" />
-    <Content Include="fonts\amcun1.flf" />
-    <Content Include="fonts\ANSI Regular.flf" />
-    <Content Include="fonts\ANSI Shadow.flf" />
-    <Content Include="fonts\Arrows.flf" />
-    <Content Include="fonts\ASCII New Roman.flf" />
-    <Content Include="fonts\ascii12.tlf" />
-    <Content Include="fonts\ascii9.tlf" />
-    <Content Include="fonts\ascii_new_roman.flf" />
-    <Content Include="fonts\Avatar.flf" />
-    <Content Include="fonts\B1FF.flf" />
-    <Content Include="fonts\B1FF.php" />
-    <Content Include="fonts\Banner.flf" />
-    <Content Include="fonts\Banner3-D.flf" />
-    <Content Include="fonts\Banner3.flf" />
-    <Content Include="fonts\Banner4.flf" />
-    <Content Include="fonts\Barbwire.flf" />
-    <Content Include="fonts\Basic.flf" />
-    <Content Include="fonts\Bear.flf" />
-    <Content Include="fonts\Bell.flf" />
-    <Content Include="fonts\Benjamin.flf" />
-    <Content Include="fonts\Big Chief.flf" />
-    <Content Include="fonts\Big Money-ne.flf" />
-    <Content Include="fonts\Big Money-nw.flf" />
-    <Content Include="fonts\Big Money-se.flf" />
-    <Content Include="fonts\Big Money-sw.flf" />
-    <Content Include="fonts\Big.flf" />
-    <Content Include="fonts\bigascii12.tlf" />
-    <Content Include="fonts\bigascii9.tlf" />
-    <Content Include="fonts\bigchief.flf" />
-    <Content Include="fonts\Bigfig.flf" />
-    <Content Include="fonts\bigmono12.tlf" />
-    <Content Include="fonts\bigmono9.tlf" />
-    <Content Include="fonts\Binary.flf" />
-    <Content Include="fonts\Block.flf" />
-    <Content Include="fonts\Blocks.flf" />
-    <Content Include="fonts\Bloody.flf" />
-    <Content Include="fonts\Bolger.flf" />
-    <Content Include="fonts\Braced.flf" />
-    <Content Include="fonts\Bright.flf" />
-    <Content Include="fonts\Broadway KB.flf" />
-    <Content Include="fonts\Broadway.flf" />
-    <Content Include="fonts\broadway_kb.flf" />
-    <Content Include="fonts\Bubble.flf" />
-    <Content Include="fonts\Bulbhead.flf" />
-    <Content Include="fonts\calgphy2.flf" />
-    <Content Include="fonts\Caligraphy.flf" />
-    <Content Include="fonts\Caligraphy2.flf" />
-    <Content Include="fonts\Calvin S.flf" />
-    <Content Include="fonts\Cards.flf" />
-    <Content Include="fonts\Catwalk.flf" />
-    <Content Include="fonts\Chiseled.flf" />
-    <Content Include="fonts\Chunky.flf" />
-    <Content Include="fonts\circle.tlf" />
-    <Content Include="fonts\Coinstak.flf" />
-    <Content Include="fonts\Cola.flf" />
-    <Content Include="fonts\Colossal.flf" />
-    <Content Include="fonts\Computer.flf" />
-    <Content Include="fonts\Contessa.flf" />
-    <Content Include="fonts\Contrast.flf" />
-    <Content Include="fonts\cosmic.flf" />
-    <Content Include="fonts\Cosmike.flf" />
-    <Content Include="fonts\Crawford.flf" />
-    <Content Include="fonts\Crawford2.flf" />
-    <Content Include="fonts\Crazy.flf" />
-    <Content Include="fonts\Cricket.flf" />
-    <Content Include="fonts\Cursive.flf" />
-    <Content Include="fonts\Cyberlarge.flf" />
-    <Content Include="fonts\Cybermedium.flf" />
-    <Content Include="fonts\Cybersmall.flf" />
-    <Content Include="fonts\Cygnet.flf" />
-    <Content Include="fonts\DANC4.flf" />
-    <Content Include="fonts\Dancing Font.flf" />
-    <Content Include="fonts\dancingfont.flf" />
-    <Content Include="fonts\Decimal.flf" />
-    <Content Include="fonts\Def Leppard.flf" />
-    <Content Include="fonts\defleppard.flf" />
-    <Content Include="fonts\Delta Corps Priest 1.flf" />
-    <Content Include="fonts\Diamond.flf" />
-    <Content Include="fonts\Diet Cola.flf" />
-    <Content Include="fonts\dietcola.flf" />
-    <Content Include="fonts\Digital.flf" />
-    <Content Include="fonts\Doh.flf" />
-    <Content Include="fonts\Doom.flf" />
-    <Content Include="fonts\DOS Rebel.flf" />
-    <Content Include="fonts\dosrebel.flf" />
-    <Content Include="fonts\Dot Matrix.flf" />
-    <Content Include="fonts\dotmatrix.flf" />
-    <Content Include="fonts\Double Shorts.flf" />
-    <Content Include="fonts\Double.flf" />
-    <Content Include="fonts\doubleshorts.flf" />
-    <Content Include="fonts\Dr Pepper.flf" />
-    <Content Include="fonts\drpepper.flf" />
-    <Content Include="fonts\DWhistled.flf" />
-    <Content Include="fonts\Efti Chess.flf" />
-    <Content Include="fonts\Efti Font.flf" />
-    <Content Include="fonts\Efti Italic.flf" />
-    <Content Include="fonts\Efti Piti.flf" />
-    <Content Include="fonts\Efti Robot.flf" />
-    <Content Include="fonts\Efti Wall.flf" />
-    <Content Include="fonts\Efti Water.flf" />
-    <Content Include="fonts\eftichess.flf" />
-    <Content Include="fonts\eftifont.flf" />
-    <Content Include="fonts\eftipiti.flf" />
-    <Content Include="fonts\eftirobot.flf" />
-    <Content Include="fonts\eftitalic.flf" />
-    <Content Include="fonts\eftiwall.flf" />
-    <Content Include="fonts\eftiwater.flf" />
-    <Content Include="fonts\Electronic.flf" />
-    <Content Include="fonts\Elite.flf" />
-    <Content Include="fonts\emboss.tlf" />
-    <Content Include="fonts\emboss2.tlf" />
-    <Content Include="fonts\Epic.flf" />
-    <Content Include="fonts\Fender.flf" />
-    <Content Include="fonts\Filter.flf" />
-    <Content Include="fonts\Fire Font-k.flf" />
-    <Content Include="fonts\Fire Font-s.flf" />
-    <Content Include="fonts\fire_font-k.flf" />
-    <Content Include="fonts\fire_font-s.flf" />
-    <Content Include="fonts\Flipped.flf" />
-    <Content Include="fonts\Flower Power.flf" />
-    <Content Include="fonts\flowerpower.flf" />
-    <Content Include="fonts\Four Tops.flf" />
-    <Content Include="fonts\fourtops.flf" />
-    <Content Include="fonts\Fraktur.flf" />
-    <Content Include="fonts\frango.flc" />
-    <Content Include="fonts\Fun Face.flf" />
-    <Content Include="fonts\Fun Faces.flf" />
-    <Content Include="fonts\funface.flf" />
-    <Content Include="fonts\funfaces.flf" />
-    <Content Include="fonts\future.tlf" />
-    <Content Include="fonts\Fuzzy.flf" />
-    <Content Include="fonts\Georgi16.flf" />
-    <Content Include="fonts\Georgia11.flf" />
-    <Content Include="fonts\Ghost.flf" />
-    <Content Include="fonts\Ghoulish.flf" />
-    <Content Include="fonts\Glenyn.flf" />
-    <Content Include="fonts\Goofy.flf" />
-    <Content Include="fonts\Gothic.flf" />
-    <Content Include="fonts\Graceful.flf" />
-    <Content Include="fonts\Gradient.flf" />
-    <Content Include="fonts\Graffiti.flf" />
-    <Content Include="fonts\Greek.flf" />
-    <Content Include="fonts\Heart Left.flf" />
-    <Content Include="fonts\Heart Right.flf" />
-    <Content Include="fonts\heart_left.flf" />
-    <Content Include="fonts\heart_right.flf" />
-    <Content Include="fonts\Henry 3D.flf" />
-    <Content Include="fonts\henry3d.flf" />
-    <Content Include="fonts\Hex.flf" />
-    <Content Include="fonts\Hieroglyphs.flf" />
-    <Content Include="fonts\Hollywood.flf" />
-    <Content Include="fonts\Horizontal Left.flf" />
-    <Content Include="fonts\Horizontal Right.flf" />
-    <Content Include="fonts\horizontalleft.flf" />
-    <Content Include="fonts\horizontalright.flf" />
-    <Content Include="fonts\hz.flc" />
-    <Content Include="fonts\ICL-1900.flf" />
-    <Content Include="fonts\ilhebrew.flc" />
-    <Content Include="fonts\Impossible.flf" />
-    <Content Include="fonts\Invita.flf" />
-    <Content Include="fonts\Isometric1.flf" />
-    <Content Include="fonts\Isometric2.flf" />
-    <Content Include="fonts\Isometric3.flf" />
-    <Content Include="fonts\Isometric4.flf" />
-    <Content Include="fonts\Italic.flf" />
-    <Content Include="fonts\Ivrit.flf" />
-    <Content Include="fonts\Jacky.flf" />
-    <Content Include="fonts\Jazmine.flf" />
-    <Content Include="fonts\Jerusalem.flf" />
-    <Content Include="fonts\jis0201.flc" />
-    <Content Include="fonts\JS Block Letters.flf" />
-    <Content Include="fonts\JS Bracket Letters.flf" />
-    <Content Include="fonts\JS Capital Curves.flf" />
-    <Content Include="fonts\JS Cursive.flf" />
-    <Content Include="fonts\JS Stick Letters.flf" />
-    <Content Include="fonts\Katakana.flf" />
-    <Content Include="fonts\Kban.flf" />
-    <Content Include="fonts\Keyboard.flf" />
-    <Content Include="fonts\Knob.flf" />
-    <Content Include="fonts\koi8r.flc" />
-    <Content Include="fonts\Konto Slant.flf" />
-    <Content Include="fonts\Konto.flf" />
-    <Content Include="fonts\kontoslant.flf" />
-    <Content Include="fonts\Larry 3D 2.flf" />
-    <Content Include="fonts\Larry 3D.flf">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-    <Content Include="fonts\larry3d.flf">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-    <Content Include="fonts\LCD.flf" />
-    <Content Include="fonts\Lean.flf" />
-    <Content Include="fonts\letter.tlf" />
-    <Content Include="fonts\Letters.flf" />
-    <Content Include="fonts\Lil Devil.flf" />
-    <Content Include="fonts\lildevil.flf" />
-    <Content Include="fonts\Line Blocks.flf" />
-    <Content Include="fonts\lineblocks.flf" />
-    <Content Include="fonts\Linux.flf" />
-    <Content Include="fonts\Lockergnome.flf" />
-    <Content Include="fonts\Madrid.flf" />
-    <Content Include="fonts\Marquee.flf" />
-    <Content Include="fonts\Maxfour.flf" />
-    <Content Include="fonts\maxiwi.flf" />
-    <Content Include="fonts\Merlin1.flf" />
-    <Content Include="fonts\Merlin2.flf" />
-    <Content Include="fonts\Mike.flf" />
-    <Content Include="fonts\Mini.flf" />
-    <Content Include="fonts\miniwi.flf" />
-    <Content Include="fonts\Mirror.flf" />
-    <Content Include="fonts\Mnemonic.flf" />
-    <Content Include="fonts\Modular.flf" />
-    <Content Include="fonts\mono12.tlf" />
-    <Content Include="fonts\mono9.tlf" />
-    <Content Include="fonts\Morse.flf" />
-    <Content Include="fonts\Morse2.flf" />
-    <Content Include="fonts\moscow.flc" />
-    <Content Include="fonts\Moscow.flf" />
-    <Content Include="fonts\Mshebrew210.flf" />
-    <Content Include="fonts\Muzzle.flf" />
-    <Content Include="fonts\Nancyj-Fancy.flf" />
-    <Content Include="fonts\Nancyj-Improved.flf" />
-    <Content Include="fonts\Nancyj-Underlined.flf" />
-    <Content Include="fonts\Nancyj.flf" />
-    <Content Include="fonts\Nipples.flf" />
-    <Content Include="fonts\NScript.flf" />
-    <Content Include="fonts\NT Greek.flf" />
-    <Content Include="fonts\ntgreek.flf" />
-    <Content Include="fonts\NV Script.flf" />
-    <Content Include="fonts\O8.flf" />
-    <Content Include="fonts\Octal.flf" />
-    <Content Include="fonts\Ogre.flf" />
-    <Content Include="fonts\Old Banner.flf" />
-    <Content Include="fonts\oldbanner.flf" />
-    <Content Include="fonts\OS2.flf" />
-    <Content Include="fonts\pagga.tlf" />
-    <Content Include="fonts\Patorjk's Cheese.flf" />
-    <Content Include="fonts\Patorjk-HeX.flf" />
-    <Content Include="fonts\Pawp.flf" />
-    <Content Include="fonts\Peaks Slant.flf" />
-    <Content Include="fonts\Peaks.flf" />
-    <Content Include="fonts\peaksslant.flf" />
-    <Content Include="fonts\Pebbles.flf" />
-    <Content Include="fonts\Pepper.flf" />
-    <Content Include="fonts\Poison.flf" />
-    <Content Include="fonts\Puffy.flf" />
-    <Content Include="fonts\Puzzle.flf" />
-    <Content Include="fonts\Pyramid.flf" />
-    <Content Include="fonts\Rammstein.flf" />
-    <Content Include="fonts\README.md" />
-    <Content Include="fonts\rebel.tlf" />
-    <Content Include="fonts\Rectangles.flf" />
-    <Content Include="fonts\Red Phoenix.flf" />
-    <Content Include="fonts\red_phoenix.flf" />
-    <Content Include="fonts\Relief.flf" />
-    <Content Include="fonts\Relief2.flf" />
-    <Content Include="fonts\rev.flf" />
-    <Content Include="fonts\Reverse.flf" />
-    <Content Include="fonts\Roman.flf" />
-    <Content Include="fonts\Rot13.flf" />
-    <Content Include="fonts\Rot13.php" />
-    <Content Include="fonts\Rotated.flf" />
-    <Content Include="fonts\Rounded.flf" />
-    <Content Include="fonts\Rowan Cap.flf" />
-    <Content Include="fonts\rowancap.flf" />
-    <Content Include="fonts\Rozzo.flf" />
-    <Content Include="fonts\Runic.flf" />
-    <Content Include="fonts\Runyc.flf" />
-    <Content Include="fonts\rusto.tlf" />
-    <Content Include="fonts\rustofat.tlf" />
-    <Content Include="fonts\S Blood.flf" />
-    <Content Include="fonts\s-relief.flf" />
-    <Content Include="fonts\Santa Clara.flf" />
-    <Content Include="fonts\santaclara.flf" />
-    <Content Include="fonts\sblood.flf" />
-    <Content Include="fonts\Script.flf" />
-    <Content Include="fonts\Serifcap.flf" />
-    <Content Include="fonts\Shadow.flf" />
-    <Content Include="fonts\Shimrod.flf" />
-    <Content Include="fonts\Short.flf" />
-    <Content Include="fonts\SL Script.flf" />
-    <Content Include="fonts\Slant Relief.flf" />
-    <Content Include="fonts\Slant.flf" />
-    <Content Include="fonts\Slide.flf" />
-    <Content Include="fonts\slscript.flf" />
-    <Content Include="fonts\Small Caps.flf" />
-    <Content Include="fonts\Small Isometric1.flf" />
-    <Content Include="fonts\Small Keyboard.flf" />
-    <Content Include="fonts\Small Poison.flf" />
-    <Content Include="fonts\Small Script.flf" />
-    <Content Include="fonts\Small Shadow.flf" />
-    <Content Include="fonts\Small Slant.flf" />
-    <Content Include="fonts\Small Tengwar.flf" />
-    <Content Include="fonts\Small.flf" />
-    <Content Include="fonts\smallcaps.flf" />
-    <Content Include="fonts\smascii12.tlf" />
-    <Content Include="fonts\smascii9.tlf" />
-    <Content Include="fonts\smblock.tlf" />
-    <Content Include="fonts\smbraille.tlf" />
-    <Content Include="fonts\smisome1.flf" />
-    <Content Include="fonts\smkeyboard.flf" />
-    <Content Include="fonts\smmono12.tlf" />
-    <Content Include="fonts\smmono9.tlf" />
-    <Content Include="fonts\smpoison.flf" />
-    <Content Include="fonts\smscript.flf" />
-    <Content Include="fonts\smshadow.flf" />
-    <Content Include="fonts\smslant.flf" />
-    <Content Include="fonts\smtengwar.flf" />
-    <Content Include="fonts\Soft.flf" />
-    <Content Include="fonts\Speed.flf" />
-    <Content Include="fonts\Spliff.flf" />
-    <Content Include="fonts\Stacey.flf" />
-    <Content Include="fonts\Stampate.flf" />
-    <Content Include="fonts\Stampatello.flf" />
-    <Content Include="fonts\Standard.flf" />
-    <Content Include="fonts\Star Strips.flf" />
-    <Content Include="fonts\Star Wars.flf" />
-    <Content Include="fonts\starstrips.flf" />
-    <Content Include="fonts\starwars.flf" />
-    <Content Include="fonts\Stellar.flf" />
-    <Content Include="fonts\Stforek.flf" />
-    <Content Include="fonts\Stick Letters.flf" />
-    <Content Include="fonts\Stop.flf" />
-    <Content Include="fonts\Straight.flf" />
-    <Content Include="fonts\Stronger Than All.flf" />
-    <Content Include="fonts\Sub-Zero.flf" />
-    <Content Include="fonts\Swamp Land.flf" />
-    <Content Include="fonts\swampland.flf" />
-    <Content Include="fonts\Swan.flf" />
-    <Content Include="fonts\Sweet.flf" />
-    <Content Include="fonts\Tanja.flf" />
-    <Content Include="fonts\Tengwar.flf" />
-    <Content Include="fonts\Term.flf" />
-    <Content Include="fonts\Test1.flf" />
-    <Content Include="fonts\The Edge.flf" />
-    <Content Include="fonts\Thick.flf" />
-    <Content Include="fonts\Thin.flf" />
-    <Content Include="fonts\THIS.flf" />
-    <Content Include="fonts\Thorned.flf" />
-    <Content Include="fonts\Three Point.flf" />
-    <Content Include="fonts\threepoint.flf" />
-    <Content Include="fonts\Ticks Slant.flf" />
-    <Content Include="fonts\Ticks.flf" />
-    <Content Include="fonts\ticksslant.flf" />
-    <Content Include="fonts\Tiles.flf" />
-    <Content Include="fonts\Tinker-Toy.flf" />
-    <Content Include="fonts\Tombstone.flf" />
-    <Content Include="fonts\Train.flf" />
-    <Content Include="fonts\Trek.flf" />
-    <Content Include="fonts\Tsalagi.flf" />
-    <Content Include="fonts\Tubular.flf" />
-    <Content Include="fonts\Twisted.flf" />
-    <Content Include="fonts\Two Point.flf" />
-    <Content Include="fonts\twopoint.flf" />
-    <Content Include="fonts\Univers.flf" />
-    <Content Include="fonts\upper.flc" />
-    <Content Include="fonts\USA Flag.flf" />
-    <Content Include="fonts\usaflag.flf" />
-    <Content Include="fonts\ushebrew.flc" />
-    <Content Include="fonts\uskata.flc" />
-    <Content Include="fonts\utf8.flc" />
-    <Content Include="fonts\Varsity.flf" />
-    <Content Include="fonts\Wavy.flf" />
-    <Content Include="fonts\Weird.flf" />
-    <Content Include="fonts\Wet Letter.flf" />
-    <Content Include="fonts\wetletter.flf" />
-    <Content Include="fonts\Whimsy.flf" />
-    <Content Include="fonts\wideterm.tlf" />
-    <Content Include="fonts\Wow.flf">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-    <Content Include="images\FinalLogo.jpg">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-    <Content Include="images\star1.png">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-    <Content Include="images\star2.jpg">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-    <Content Include="images\star3.png">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-    <Content Include="images\star4.png">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-    <Content Include="images\star5.png">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-    <Content Include="images\star6.jpg">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-    <Content Include="images\star6b.jpg">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-    <Content Include="Lib\Spectre.Console.deps.json">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-    <Content Include="Lib\Spectre.Console.xml">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-    <Content Include="Spectre.Console.deps.json">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-    <Content Include="Spectre.Console.xml">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\NextGenSoftware.OASIS.STAR\NextGenSoftware.OASIS.STAR.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Genesis\" />
+    <ProjectReference Include="..\NextGenSoftware.OASIS.STAR.STARDNA\NextGenSoftware.OASIS.STAR.DNA.csproj" />
   </ItemGroup>
 
 </Project>
