@@ -28,6 +28,7 @@ import {
   LibraryBooks,
   Extension,
   Inventory,
+  Person,
   LocalFireDepartment,
   Store,
   Settings,
@@ -38,7 +39,15 @@ import {
   CloudUpload,
   Code,
   DataObject,
+  AccountBalanceWallet,
+  Terminal,
+  CreditCard,
+  TrendingUp,
+  NetworkCheck,
+  Search as SearchIcon,
 } from '@mui/icons-material';
+import KeyIcon from '@mui/icons-material/VpnKey';
+import VideoCallIcon from '@mui/icons-material/VideoCall';
 import { useNavigate, useLocation } from 'react-router-dom';
 import { motion } from 'framer-motion';
 import { useAvatar } from '../contexts/AvatarContext';
@@ -100,6 +109,48 @@ const Sidebar: React.FC<SidebarProps> = ({ open, onClose, isConnected }) => {
       ],
     },
     {
+      id: 'tools',
+      title: 'Tools',
+      icon: <Settings />,
+      children: [
+        {
+          id: 'onet',
+          title: 'ONET',
+          icon: <NetworkCheck />,
+          path: '/ONET',
+          description: 'OASIS P2P Network',
+        },
+        {
+          id: 'keys',
+          title: 'Keys',
+          icon: <KeyIcon />,
+          path: '/keys',
+          description: 'Key Management',
+        },
+        {
+          id: 'search',
+          title: 'Search',
+          icon: <SearchIcon />,
+          path: '/search',
+          description: 'Global Search',
+        },
+        {
+          id: 'map',
+          title: 'Map',
+          icon: <LocationOn />,
+          path: '/map',
+          description: 'Geo Map',
+        },
+        {
+          id: 'hypernet',
+          title: 'HyperNET',
+          icon: <Memory />,
+          path: '/hypernet',
+          description: 'HyperNET Management',
+        },
+      ],
+    },
+    {
       id: 'nfts',
       title: 'NFTs & Assets',
       icon: <Image />,
@@ -110,6 +161,13 @@ const Sidebar: React.FC<SidebarProps> = ({ open, onClose, isConnected }) => {
           icon: <Image />,
           path: '/nfts',
           description: 'Digital Assets',
+        },
+        {
+          id: 'nft-minting',
+          title: 'NFT Minting',
+          icon: <CloudUpload />,
+          path: '/nfts/mint',
+          description: 'Mint New NFTs',
         },
         {
           id: 'geonfts',
@@ -156,6 +214,41 @@ const Sidebar: React.FC<SidebarProps> = ({ open, onClose, isConnected }) => {
       ],
     },
     {
+      id: 'my-area',
+      title: 'My Area',
+      icon: <Person />,
+      children: [
+        {
+          id: 'wallets',
+          title: 'Wallets',
+          icon: <AccountBalanceWallet />,
+          path: '/wallets',
+          description: 'OASIS Universal Wallet',
+        },
+        {
+          id: 'subscription-plans',
+          title: 'Subscription Plans',
+          icon: <CreditCard />,
+          path: '/subscription/plans',
+          description: 'OASIS Subscription Plans',
+        },
+        {
+          id: 'my-data',
+          title: 'My Data',
+          icon: <CloudUpload />,
+          path: '/my-data',
+          description: 'OASIS Hyperdrive',
+        },
+        {
+          id: 'usage',
+          title: 'Usage & Billing',
+          icon: <TrendingUp />,
+          path: '/subscription/usage',
+          description: 'API Usage & Billing',
+        },
+      ],
+    },
+    {
       id: 'development',
       title: 'Development',
       icon: <Memory />,
@@ -189,6 +282,13 @@ const Sidebar: React.FC<SidebarProps> = ({ open, onClose, isConnected }) => {
           description: 'OASIS Data Objects',
         },
         {
+          id: 'metadata',
+          title: 'MetaData',
+          icon: <DataObject />,
+          path: '/metadata',
+          description: 'Holon Metadata',
+        },
+        {
           id: 'zomes',
           title: 'Zomes',
           icon: <Extension />,
@@ -201,6 +301,34 @@ const Sidebar: React.FC<SidebarProps> = ({ open, onClose, isConnected }) => {
           icon: <Extension />,
           path: '/plugins',
           description: 'Extensions & Plugins',
+        },
+        {
+          id: 'star-cli',
+          title: 'STAR CLI',
+          icon: <Terminal />,
+          path: '/star-cli',
+          description: 'Command Line Interface',
+        },
+        {
+          id: 'oasisdna',
+          title: 'OASISDNA Config',
+          icon: <Settings />,
+          path: '/oasisdna',
+          description: 'OASIS DNA Configuration',
+        },
+        {
+          id: 'onet',
+          title: 'ONET P2P',
+          icon: <NetworkCheck />,
+          path: '/onet',
+          description: 'OASIS P2P Network',
+        },
+        {
+          id: 'onode',
+          title: 'ONODE',
+          icon: <Memory />,
+          path: '/onode',
+          description: 'OASIS Node Management',
         },
       ],
     },
@@ -231,11 +359,46 @@ const Sidebar: React.FC<SidebarProps> = ({ open, onClose, isConnected }) => {
           description: 'Karma Leaderboard',
         },
         {
-          id: 'my-data',
-          title: 'My Data',
-          icon: <CloudUpload />,
-          path: '/my-data',
-          description: 'OASIS Hyperdrive',
+          id: 'competition',
+          title: 'Competition',
+          icon: <EmojiEvents />,
+          path: '/competition',
+          description: 'Seasons, Ranks & Leagues',
+        },
+        {
+          id: 'messaging',
+          title: 'Messaging',
+          icon: <AccountCircle />,
+          path: '/messaging',
+          description: 'Direct Messages',
+        },
+        {
+          id: 'chat',
+          title: 'Chat',
+          icon: <AccountCircle />,
+          path: '/chat',
+          description: 'Channels & Group Chat',
+        },
+        {
+          id: 'social',
+          title: 'Social',
+          icon: <AccountCircle />,
+          path: '/social',
+          description: 'Feed & Sharing',
+        },
+        {
+          id: 'eggs',
+          title: 'Eggs',
+          icon: <Image />,
+          path: '/eggs',
+          description: 'Discover & Hatch',
+        },
+        {
+          id: 'video',
+          title: 'Video Calls',
+          icon: <VideoCallIcon />,
+          path: '/video',
+          description: 'Start & Manage Calls',
         },
       ],
     },
