@@ -20,6 +20,7 @@ import {
   ListItemAvatar,
   ListItemSecondaryAction,
   Badge,
+  Button,
 } from '@mui/material';
 import {
   TrendingUp,
@@ -45,18 +46,46 @@ import {
   Store,
   Science,
   Explore,
+  Help,
+  Apps,
+  Memory,
+  Description,
+  ShoppingCart,
+  Visibility,
+  Download,
+  Upload,
+  PlayArrow,
+  Pause,
 } from '@mui/icons-material';
 import { motion } from 'framer-motion';
 import { useQuery } from 'react-query';
-import { starService } from '../services/starService';
+import { starCoreService, avatarService } from '../services';
 import { LineChart, Line, XAxis, YAxis, CartesianGrid, Tooltip as RechartsTooltip, ResponsiveContainer, BarChart, Bar, PieChart, Pie, Cell } from 'recharts';
+import { useNavigate } from 'react-router-dom';
+import CloseIcon from '@mui/icons-material/Close';
 
 interface DashboardProps {
   isConnected: boolean;
 }
 
 const Dashboard: React.FC<DashboardProps> = ({ isConnected }) => {
+  const [showInfo, setShowInfo] = useState(true);
   const [refreshKey, setRefreshKey] = useState(0);
+  const navigate = useNavigate();
+  const DismissibleInfoBar: React.FC = () => {
+    if (!showInfo) return null;
+    return (
+      <Box sx={{ mt: 1, p: 2, bgcolor: '#0d47a1', color: 'white', borderRadius: 1, display: 'flex', alignItems: 'center', gap: 1, position: 'relative' }}>
+        <Info sx={{ color: 'white' }} />
+        <Typography variant="body2" sx={{ color: 'white' }}>
+          <strong>System Status:</strong> Monitor OAPPs, Runtimes, Templates, and all ecosystem components. Click refresh to update real-time data.
+        </Typography>
+        <IconButton size="small" onClick={() => setShowInfo(false)} sx={{ position: 'absolute', right: 8, top: 8, color: 'white' }}>
+          <CloseIcon />
+        </IconButton>
+      </Box>
+    );
+  };
 
   // Fetch dashboard data
   const { data: dashboardData, isLoading, error, refetch } = useQuery(
@@ -64,7 +93,7 @@ const Dashboard: React.FC<DashboardProps> = ({ isConnected }) => {
     async () => {
       try {
         // Try to get real data first
-        const response = await starService.getDashboardData?.();
+        const response = await starCoreService.getDashboardData?.();
         return response;
       } catch (error) {
         // Fallback to impressive demo data
@@ -98,13 +127,54 @@ const Dashboard: React.FC<DashboardProps> = ({ isConnected }) => {
             },
             recentActivity: [
               { id: 1, type: 'user', action: 'New user registered', user: 'John Doe', time: '2 minutes ago', status: 'success' },
-              { id: 2, type: 'oapp', action: 'OAPP deployed', name: 'Quantum Calculator', time: '5 minutes ago', status: 'success' },
-              { id: 3, type: 'nft', action: 'NFT minted', name: 'Cosmic Dragon', time: '8 minutes ago', status: 'success' },
-              { id: 4, type: 'transaction', action: 'Payment processed', amount: '$2,500', time: '12 minutes ago', status: 'success' },
-              { id: 5, type: 'error', action: 'System warning', message: 'High CPU usage detected', time: '15 minutes ago', status: 'warning' },
-              { id: 6, type: 'avatar', action: 'Avatar created', name: 'Space Explorer', time: '18 minutes ago', status: 'success' },
-              { id: 7, type: 'runtime', action: 'Runtime updated', name: 'Node.js 18', time: '22 minutes ago', status: 'success' },
-              { id: 8, type: 'library', action: 'Library published', name: 'AI Toolkit', time: '25 minutes ago', status: 'success' },
+              { id: 2, type: 'oapp', action: 'OAPP published', user: 'Sarah Wilson', time: '5 minutes ago', status: 'success' },
+              { id: 3, type: 'runtime', action: 'Runtime activated', user: 'Mike Chen', time: '8 minutes ago', status: 'success' },
+              { id: 4, type: 'template', action: 'Template downloaded', user: 'Emma Davis', time: '12 minutes ago', status: 'success' },
+              { id: 5, type: 'store', action: 'Asset purchased', user: 'Alex Brown', time: '15 minutes ago', status: 'success' },
+            ],
+            topOAPPs: [
+              { id: '1', name: 'Cosmic Explorer', downloads: 15420, rating: 4.8, author: 'SpaceDev Studios', category: 'Exploration' },
+              { id: '2', name: 'Quantum Builder', downloads: 8930, rating: 4.9, author: 'Quantum Labs', category: 'Construction' },
+              { id: '3', name: 'Neural Network Manager', downloads: 25670, rating: 4.7, author: 'AI Innovations', category: 'AI/ML' },
+            ],
+            topRuntimes: [
+              { id: '1', name: 'JavaScript Engine', active: 189, total: 234, type: 'Programming Language', status: 'Running' },
+              { id: '2', name: 'Python Interpreter', active: 156, total: 189, type: 'Programming Language', status: 'Running' },
+              { id: '3', name: 'Node.js Runtime', active: 134, total: 167, type: 'Web Runtime', status: 'Running' },
+            ],
+            topTemplates: [
+              { id: '1', name: 'React SPA Template', downloads: 234000, rating: 4.8, author: 'React Team', category: 'Web App' },
+              { id: '2', name: 'Vue.js SPA Template', downloads: 89000, rating: 4.7, author: 'Vue Team', category: 'Web App' },
+              { id: '3', name: 'Angular Enterprise', downloads: 67000, rating: 4.6, author: 'Microsoft', category: 'Web App' },
+            ],
+            storeItems: [
+              { id: '1', name: 'Quantum Engine Core', price: 25000, sales: 1247, category: 'Hardware', rating: 4.9 },
+              { id: '2', name: 'Neural Network SDK', price: 8500, sales: 892, category: 'Software', rating: 4.8 },
+              { id: '3', name: 'Holographic Display Array', price: 45000, sales: 156, category: 'Display', rating: 4.7 },
+            ],
+            systemMetrics: {
+              cpuUsage: 45.2,
+              memoryUsage: 67.8,
+              diskUsage: 34.5,
+              networkLatency: 12.3,
+              activeConnections: 892,
+              totalRequests: 4567892,
+            },
+            userGrowth: [
+              { month: 'Jan', users: 1200000, karma: 5000000 },
+              { month: 'Feb', users: 1350000, karma: 6000000 },
+              { month: 'Mar', users: 1500000, karma: 7500000 },
+              { month: 'Apr', users: 1700000, karma: 8500000 },
+              { month: 'May', users: 1900000, karma: 9500000 },
+              { month: 'Jun', users: 2200000, karma: 11000000 },
+              { month: 'Jul', users: 2547891, karma: 12500000 },
+            ],
+            categoryDistribution: [
+              { name: 'Web Apps', value: 35, count: 1250 },
+              { name: 'Mobile Apps', value: 25, count: 892 },
+              { name: 'AI/ML', value: 20, count: 714 },
+              { name: 'Games', value: 15, count: 536 },
+              { name: 'Blockchain', value: 5, count: 179 },
             ],
             performanceData: [
               { name: 'Jan', users: 1200000, karma: 2100000, transactions: 45000 },
@@ -204,12 +274,20 @@ const Dashboard: React.FC<DashboardProps> = ({ isConnected }) => {
       <Box sx={{ mt: 4, mb: 4, overflow: 'visible' }}>
         <Box sx={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', mb: 3, px: { xs: 0, md: 1 } }}>
           <Box>
-            <Typography variant="h4" gutterBottom className="page-heading">
-              📊 Dashboard
-            </Typography>
+            <Box sx={{ display: 'flex', alignItems: 'center', gap: 2, mb: 1 }}>
+              <Typography variant="h4" gutterBottom className="page-heading">
+                📊 Dashboard
+              </Typography>
+              <Tooltip title="Dashboard provides real-time analytics and system overview. Monitor your OASIS ecosystem health, user activity, and system performance.">
+                <IconButton size="small" color="primary">
+                  <Help />
+                </IconButton>
+              </Tooltip>
+            </Box>
             <Typography variant="subtitle1" color="text.secondary">
               Real-time analytics and system overview
             </Typography>
+            <DismissibleInfoBar />
           </Box>
           <Box sx={{ display: 'flex', gap: 2 }}>
             <Tooltip title="Refresh Data">
@@ -505,6 +583,413 @@ const Dashboard: React.FC<DashboardProps> = ({ isConnected }) => {
                       </Grid>
                     ))}
                   </Grid>
+                </CardContent>
+              </Card>
+            </motion.div>
+          </Grid>
+        </Grid>
+
+        {/* Interactive Dashboard Sections */}
+        <Grid container spacing={3} sx={{ mt: 2, overflow: 'visible', pb: 3, px: { xs: 0, md: 1 }, width: '100%' }}>
+          {/* Top OAPPs Section */}
+          <Grid item xs={12} md={6}>
+            <motion.div variants={itemVariants}>
+              <Card sx={{ height: '100%' }}>
+                <Box sx={{ display: 'flex', alignItems: 'center', justifyContent: 'space-between', p: 2, borderBottom: 1, borderColor: 'divider' }}>
+                  <Typography variant="h6" sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
+                    <Apps color="primary" />
+                    Top OAPPs
+                  </Typography>
+                  <Button 
+                    size="small" 
+                    onClick={() => navigate('/oapps')}
+                    endIcon={<ArrowUpward />}
+                  >
+                    View All
+                  </Button>
+                </Box>
+                <CardContent sx={{ p: 0 }}>
+                  <List>
+                    {[
+                      { id: '1', name: 'Cosmic Explorer', downloads: 15420, rating: 4.8, author: 'SpaceDev Studios', category: 'Exploration' },
+                      { id: '2', name: 'Quantum Builder', downloads: 8930, rating: 4.9, author: 'Quantum Labs', category: 'Construction' },
+                      { id: '3', name: 'Neural Network Manager', downloads: 25670, rating: 4.7, author: 'AI Innovations', category: 'AI/ML' },
+                    ].map((oapp) => (
+                      <ListItem 
+                        key={oapp.id} 
+                        sx={{ 
+                          cursor: 'pointer', 
+                          '&:hover': { bgcolor: 'action.hover' },
+                          borderRadius: 1,
+                          mb: 1,
+                          mx: 1
+                        }}
+                        onClick={() => navigate(`/oapps/${oapp.id}`)}
+                      >
+                        <ListItemAvatar>
+                          <Avatar sx={{ bgcolor: 'primary.main' }}>
+                            <Apps />
+                          </Avatar>
+                        </ListItemAvatar>
+                        <ListItemText
+                          primary={oapp.name}
+                          secondary={
+                            <Box>
+                              <Typography variant="body2" color="text.secondary">
+                                {oapp.author} • {oapp.category}
+                              </Typography>
+                              <Box sx={{ display: 'flex', alignItems: 'center', gap: 2, mt: 0.5 }}>
+                                <Chip 
+                                  icon={<Download />} 
+                                  label={`${oapp.downloads.toLocaleString()} downloads`} 
+                                  size="small" 
+                                  color="primary" 
+                                  variant="outlined"
+                                />
+                                <Chip 
+                                  icon={<Star />} 
+                                  label={`${oapp.rating}/5`} 
+                                  size="small" 
+                                  color="secondary" 
+                                  variant="outlined"
+                                />
+                              </Box>
+                            </Box>
+                          }
+                        />
+                        <ListItemSecondaryAction>
+                          <IconButton size="small" onClick={() => navigate(`/oapps/${oapp.id}`)}>
+                            <Visibility />
+                          </IconButton>
+                        </ListItemSecondaryAction>
+                      </ListItem>
+                    ))}
+                  </List>
+                </CardContent>
+              </Card>
+            </motion.div>
+          </Grid>
+
+          {/* Top Runtimes Section */}
+          <Grid item xs={12} md={6}>
+            <motion.div variants={itemVariants}>
+              <Card sx={{ height: '100%' }}>
+                <Box sx={{ display: 'flex', alignItems: 'center', justifyContent: 'space-between', p: 2, borderBottom: 1, borderColor: 'divider' }}>
+                  <Typography variant="h6" sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
+                    <Memory color="secondary" />
+                    Top Runtimes
+                  </Typography>
+                  <Button 
+                    size="small" 
+                    onClick={() => navigate('/runtimes')}
+                    endIcon={<ArrowUpward />}
+                  >
+                    View All
+                  </Button>
+                </Box>
+                <CardContent sx={{ p: 0 }}>
+                  <List>
+                    {[
+                      { id: '1', name: 'JavaScript Engine', active: 189, total: 234, type: 'Programming Language', status: 'Running' },
+                      { id: '2', name: 'Python Interpreter', active: 156, total: 189, type: 'Programming Language', status: 'Running' },
+                      { id: '3', name: 'Node.js Runtime', active: 134, total: 167, type: 'Web Runtime', status: 'Running' },
+                    ].map((runtime) => (
+                      <ListItem 
+                        key={runtime.id} 
+                        sx={{ 
+                          cursor: 'pointer', 
+                          '&:hover': { bgcolor: 'action.hover' },
+                          borderRadius: 1,
+                          mb: 1,
+                          mx: 1
+                        }}
+                        onClick={() => navigate(`/runtimes/${runtime.id}`)}
+                      >
+                        <ListItemAvatar>
+                          <Avatar sx={{ bgcolor: 'secondary.main' }}>
+                            <Memory />
+                          </Avatar>
+                        </ListItemAvatar>
+                        <ListItemText
+                          primary={runtime.name}
+                          secondary={
+                            <Box>
+                              <Typography variant="body2" color="text.secondary">
+                                {runtime.type}
+                              </Typography>
+                              <Box sx={{ display: 'flex', alignItems: 'center', gap: 2, mt: 0.5 }}>
+                                <Chip 
+                                  icon={runtime.status === 'Running' ? <PlayArrow /> : <Pause />} 
+                                  label={`${runtime.active}/${runtime.total} active`} 
+                                  size="small" 
+                                  color={runtime.status === 'Running' ? 'success' : 'default'}
+                                  variant="outlined"
+                                />
+                                <Chip 
+                                  label={runtime.status} 
+                                  size="small" 
+                                  color={runtime.status === 'Running' ? 'success' : 'default'}
+                                  variant="outlined"
+                                />
+                              </Box>
+                            </Box>
+                          }
+                        />
+                        <ListItemSecondaryAction>
+                          <IconButton size="small" onClick={() => navigate(`/runtimes/${runtime.id}`)}>
+                            <Visibility />
+                          </IconButton>
+                        </ListItemSecondaryAction>
+                      </ListItem>
+                    ))}
+                  </List>
+                </CardContent>
+              </Card>
+            </motion.div>
+          </Grid>
+
+          {/* Top Templates Section */}
+          <Grid item xs={12} md={6}>
+            <motion.div variants={itemVariants}>
+              <Card sx={{ height: '100%' }}>
+                <Box sx={{ display: 'flex', alignItems: 'center', justifyContent: 'space-between', p: 2, borderBottom: 1, borderColor: 'divider' }}>
+                  <Typography variant="h6" sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
+                    <Description color="success" />
+                    Top Templates
+                  </Typography>
+                  <Button 
+                    size="small" 
+                    onClick={() => navigate('/templates')}
+                    endIcon={<ArrowUpward />}
+                  >
+                    View All
+                  </Button>
+                </Box>
+                <CardContent sx={{ p: 0 }}>
+                  <List>
+                    {[
+                      { id: '1', name: 'React SPA Template', downloads: 234000, rating: 4.8, author: 'React Team', category: 'Web App' },
+                      { id: '2', name: 'Vue.js SPA Template', downloads: 89000, rating: 4.7, author: 'Vue Team', category: 'Web App' },
+                      { id: '3', name: 'Angular Enterprise', downloads: 67000, rating: 4.6, author: 'Microsoft', category: 'Web App' },
+                    ].map((template) => (
+                      <ListItem 
+                        key={template.id} 
+                        sx={{ 
+                          cursor: 'pointer', 
+                          '&:hover': { bgcolor: 'action.hover' },
+                          borderRadius: 1,
+                          mb: 1,
+                          mx: 1
+                        }}
+                        onClick={() => navigate(`/templates/${template.id}`)}
+                      >
+                        <ListItemAvatar>
+                          <Avatar sx={{ bgcolor: 'success.main' }}>
+                            <Description />
+                          </Avatar>
+                        </ListItemAvatar>
+                        <ListItemText
+                          primary={template.name}
+                          secondary={
+                            <Box>
+                              <Typography variant="body2" color="text.secondary">
+                                {template.author} • {template.category}
+                              </Typography>
+                              <Box sx={{ display: 'flex', alignItems: 'center', gap: 2, mt: 0.5 }}>
+                                <Chip 
+                                  icon={<Download />} 
+                                  label={`${template.downloads.toLocaleString()} downloads`} 
+                                  size="small" 
+                                  color="primary" 
+                                  variant="outlined"
+                                />
+                                <Chip 
+                                  icon={<Star />} 
+                                  label={`${template.rating}/5`} 
+                                  size="small" 
+                                  color="secondary" 
+                                  variant="outlined"
+                                />
+                              </Box>
+                            </Box>
+                          }
+                        />
+                        <ListItemSecondaryAction>
+                          <IconButton size="small" onClick={() => navigate(`/templates/${template.id}`)}>
+                            <Visibility />
+                          </IconButton>
+                        </ListItemSecondaryAction>
+                      </ListItem>
+                    ))}
+                  </List>
+                </CardContent>
+              </Card>
+            </motion.div>
+          </Grid>
+
+          {/* Store Items Section */}
+          <Grid item xs={12} md={6}>
+            <motion.div variants={itemVariants}>
+              <Card sx={{ height: '100%' }}>
+                <Box sx={{ display: 'flex', alignItems: 'center', justifyContent: 'space-between', p: 2, borderBottom: 1, borderColor: 'divider' }}>
+                  <Typography variant="h6" sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
+                    <ShoppingCart color="warning" />
+                    Top Store Items
+                  </Typography>
+                  <Button 
+                    size="small" 
+                    onClick={() => navigate('/store')}
+                    endIcon={<ArrowUpward />}
+                  >
+                    View All
+                  </Button>
+                </Box>
+                <CardContent sx={{ p: 0 }}>
+                  <List>
+                    {[
+                      { id: '1', name: 'Quantum Engine Core', price: 25000, sales: 1247, category: 'Hardware', rating: 4.9 },
+                      { id: '2', name: 'Neural Network SDK', price: 8500, sales: 892, category: 'Software', rating: 4.8 },
+                      { id: '3', name: 'Holographic Display Array', price: 45000, sales: 156, category: 'Display', rating: 4.7 },
+                    ].map((item) => (
+                      <ListItem 
+                        key={item.id} 
+                        sx={{ 
+                          cursor: 'pointer', 
+                          '&:hover': { bgcolor: 'action.hover' },
+                          borderRadius: 1,
+                          mb: 1,
+                          mx: 1
+                        }}
+                        onClick={() => navigate(`/store/${item.id}`)}
+                      >
+                        <ListItemAvatar>
+                          <Avatar sx={{ bgcolor: 'warning.main' }}>
+                            <ShoppingCart />
+                          </Avatar>
+                        </ListItemAvatar>
+                        <ListItemText
+                          primary={item.name}
+                          secondary={
+                            <Box>
+                              <Typography variant="body2" color="text.secondary">
+                                {item.category}
+                              </Typography>
+                              <Box sx={{ display: 'flex', alignItems: 'center', gap: 2, mt: 0.5 }}>
+                                <Chip 
+                                  label={`$${item.price.toLocaleString()}`} 
+                                  size="small" 
+                                  color="success" 
+                                  variant="outlined"
+                                />
+                                <Chip 
+                                  icon={<Star />} 
+                                  label={`${item.rating}/5`} 
+                                  size="small" 
+                                  color="secondary" 
+                                  variant="outlined"
+                                />
+                                <Chip 
+                                  label={`${item.sales} sales`} 
+                                  size="small" 
+                                  color="info" 
+                                  variant="outlined"
+                                />
+                              </Box>
+                            </Box>
+                          }
+                        />
+                        <ListItemSecondaryAction>
+                          <IconButton size="small" onClick={() => navigate(`/store/${item.id}`)}>
+                            <Visibility />
+                          </IconButton>
+                        </ListItemSecondaryAction>
+                      </ListItem>
+                    ))}
+                  </List>
+                </CardContent>
+              </Card>
+            </motion.div>
+          </Grid>
+        </Grid>
+
+        {/* Category Distribution Chart */}
+        <Grid container spacing={3} sx={{ mt: 2, overflow: 'visible', pb: 3, px: { xs: 0, md: 1 }, width: '100%' }}>
+          <Grid item xs={12} md={6}>
+            <motion.div variants={itemVariants}>
+              <Card sx={{ height: 400 }}>
+                <Box sx={{ p: 2, borderBottom: 1, borderColor: 'divider' }}>
+                  <Typography variant="h6" sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
+                    <Explore color="info" />
+                    Category Distribution
+                  </Typography>
+                </Box>
+                <CardContent>
+                  <ResponsiveContainer width="100%" height={300}>
+                    <PieChart>
+                      <Pie
+                        data={[
+                          { name: 'Web Apps', value: 35, count: 1250 },
+                          { name: 'Mobile Apps', value: 25, count: 892 },
+                          { name: 'AI/ML', value: 20, count: 714 },
+                          { name: 'Games', value: 15, count: 536 },
+                          { name: 'Blockchain', value: 5, count: 179 },
+                        ]}
+                        cx="50%"
+                        cy="50%"
+                        labelLine={false}
+                        label={({ name, percent }: { name: string; percent: number }) => `${name} ${(percent * 100).toFixed(0)}%`}
+                        outerRadius={80}
+                        fill="#8884d8"
+                        dataKey="value"
+                      >
+                        {[
+                          { name: 'Web Apps', value: 35, count: 1250 },
+                          { name: 'Mobile Apps', value: 25, count: 892 },
+                          { name: 'AI/ML', value: 20, count: 714 },
+                          { name: 'Games', value: 15, count: 536 },
+                          { name: 'Blockchain', value: 5, count: 179 },
+                        ].map((entry, index) => (
+                          <Cell key={`cell-${index}`} fill={['#1976d2', '#4caf50', '#ff9800', '#f44336', '#9c27b0'][index % 5]} />
+                        ))}
+                      </Pie>
+                      <RechartsTooltip />
+                    </PieChart>
+                  </ResponsiveContainer>
+                </CardContent>
+              </Card>
+            </motion.div>
+          </Grid>
+
+          {/* User Growth Chart */}
+          <Grid item xs={12} md={6}>
+            <motion.div variants={itemVariants}>
+              <Card sx={{ height: 400 }}>
+                <Box sx={{ p: 2, borderBottom: 1, borderColor: 'divider' }}>
+                  <Typography variant="h6" sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
+                    <TrendingUp color="primary" />
+                    User Growth & Karma
+                  </Typography>
+                </Box>
+                <CardContent>
+                  <ResponsiveContainer width="100%" height={300}>
+                    <LineChart data={[
+                      { month: 'Jan', users: 1200000, karma: 5000000 },
+                      { month: 'Feb', users: 1350000, karma: 6000000 },
+                      { month: 'Mar', users: 1500000, karma: 7500000 },
+                      { month: 'Apr', users: 1700000, karma: 8500000 },
+                      { month: 'May', users: 1900000, karma: 9500000 },
+                      { month: 'Jun', users: 2200000, karma: 11000000 },
+                      { month: 'Jul', users: 2547891, karma: 12500000 },
+                    ]}>
+                      <CartesianGrid strokeDasharray="3 3" />
+                      <XAxis dataKey="month" />
+                      <YAxis yAxisId="left" />
+                      <YAxis yAxisId="right" orientation="right" />
+                      <RechartsTooltip />
+                      <Line yAxisId="left" type="monotone" dataKey="users" stroke="#1976d2" strokeWidth={2} />
+                      <Line yAxisId="right" type="monotone" dataKey="karma" stroke="#4caf50" strokeWidth={2} />
+                    </LineChart>
+                  </ResponsiveContainer>
                 </CardContent>
               </Card>
             </motion.div>
