@@ -1,652 +1,638 @@
+/**
+ * GeoNFTs Page
+ * Complete GeoNFT management interface
+ */
+
 import React, { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { useDemoMode } from '../contexts/DemoModeContext';
 import {
   Box,
   Typography,
+  Button,
   Card,
   CardContent,
   Grid,
-  Button,
   Chip,
   IconButton,
+  Menu,
+  MenuItem,
   Dialog,
   DialogTitle,
   DialogContent,
   DialogActions,
   TextField,
-  Select,
-  MenuItem,
   FormControl,
   InputLabel,
-  Alert,
-  CircularProgress,
+  Select,
+  Fab,
+  Tooltip,
+  Tabs,
+  Tab,
   Badge,
+  Stack,
+  Avatar,
+  CardMedia,
+  CardActions,
+  Divider,
 } from '@mui/material';
 import {
-  LocationOn,
   Add,
-  Refresh,
-  FilterList,
-  Visibility,
+  MoreVert,
+  PlayArrow,
+  Pause,
+  Download,
+  Upload,
   Delete,
-  Public,
-  Terrain,
-  Business,
-  Nature,
+  Edit,
+  Visibility,
+  LocationOn,
+  FilterList,
+  Search,
+  Help,
+  Info,
+  Build,
+  MonetizationOn,
+  Share,
+  Favorite,
+  Star,
+  Map,
 } from '@mui/icons-material';
 import { motion } from 'framer-motion';
 import { useQuery, useMutation, useQueryClient } from 'react-query';
-import { starService } from '../services/starService';
-import toast from 'react-hot-toast';
-import { useNavigate } from 'react-router-dom';
+import { geoNftService } from '../services';
+import { toast } from 'react-hot-toast';
 
-interface GeoNFT {
-  id: string;
-  name: string;
-  description: string;
-  imageUrl: string;
-  latitude: number;
-  longitude: number;
-  location: string;
-  category: string;
-  rarity: string;
-  price: number;
-  owner: string;
-  createdAt: string;
-  isForSale: boolean;
-  views: number;
-  likes: number;
-  isInstalled?: boolean; // Added for installed badge and filtering
+interface TabPanelProps {
+  children?: React.ReactNode;
+  index: number;
+  value: number;
+}
+
+function TabPanel(props: TabPanelProps) {
+  const { children, value, index, ...other } = props;
+  return (
+    <div
+      role="tabpanel"
+      hidden={value !== index}
+      id={`geonft-tabpanel-${index}`}
+      aria-labelledby={`geonft-tab-${index}`}
+      {...other}
+    >
+      {value === index && <Box sx={{ p: 3 }}>{children}</Box>}
+    </div>
+  );
 }
 
 const GeoNFTsPage: React.FC = () => {
   const navigate = useNavigate();
+  const { isDemoMode } = useDemoMode();
+  const queryClient = useQueryClient();
+  const [tabValue, setTabValue] = useState(0);
+  const [searchTerm, setSearchTerm] = useState('');
+  const [filterType, setFilterType] = useState('all');
+  const [sortBy, setSortBy] = useState('newest');
   const [createDialogOpen, setCreateDialogOpen] = useState(false);
-  const [filterCategory, setFilterCategory] = useState('all');
-  const [viewScope, setViewScope] = useState<string>('all');
-  const [newGeoNFT, setNewGeoNFT] = useState({
-    name: '',
-    description: '',
-    imageUrl: '',
-    latitude: 0,
-    longitude: 0,
-    location: '',
-    category: 'Landmark',
-    rarity: 'Common',
-    price: 0,
+  const [selectedGeoNFT, setSelectedGeoNFT] = useState<any>(null);
+  const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
+
+  // Fetch GeoNFTs
+  const { data: geoNFTs, isLoading, error } = useQuery('geoNFTs', () => geoNftService.getAll());
+
+  // Create GeoNFT mutation
+  const createGeoNFTMutation = useMutation(
+    async (geoNFTData: any) => {
+      const response = await geoNftService.create(geoNFTData);
+      return response.result;
+    },
+    {
+      onSuccess: () => {
+        queryClient.invalidateQueries('geoNFTs');
+        toast.success('GeoNFT created successfully!');
+        setCreateDialogOpen(false);
+      },
+      onError: (error: any) => {
+        toast.error('Failed to create GeoNFT: ' + error.message);
+      },
+    }
+  );
+
+  // Publish GeoNFT mutation
+  const publishGeoNFTMutation = useMutation(
+    async (geoNFTId: string) => {
+      const response = await geoNftService.publish(geoNFTId, {});
+      return response.result;
+    },
+    {
+      onSuccess: () => {
+        queryClient.invalidateQueries('geoNFTs');
+        toast.success('GeoNFT published successfully!');
+      },
+      onError: (error: any) => {
+        toast.error('Failed to publish GeoNFT: ' + error.message);
+      },
+    }
+  );
+
+  // Download GeoNFT mutation
+  const downloadGeoNFTMutation = useMutation(
+    async (geoNFTId: string) => {
+      const response = await geoNftService.download(geoNFTId, './downloads', true);
+      return response.result;
+    },
+    {
+      onSuccess: () => {
+        toast.success('GeoNFT downloaded successfully!');
+      },
+      onError: (error: any) => {
+        toast.error('Failed to download GeoNFT: ' + error.message);
+      },
+    }
+  );
+
+  const handleCreateGeoNFT = (geoNFTData: any) => {
+    createGeoNFTMutation.mutate(geoNFTData);
+  };
+
+  const handlePublishGeoNFT = (geoNFTId: string) => {
+    publishGeoNFTMutation.mutate(geoNFTId);
+  };
+
+  const handleDownloadGeoNFT = (geoNFTId: string) => {
+    downloadGeoNFTMutation.mutate(geoNFTId);
+  };
+
+  const handleMenuClick = (event: React.MouseEvent<HTMLElement>, geoNFT: any) => {
+    setAnchorEl(event.currentTarget);
+    setSelectedGeoNFT(geoNFT);
+  };
+
+  const handleMenuClose = () => {
+    setAnchorEl(null);
+    setSelectedGeoNFT(null);
+  };
+
+  const filteredGeoNFTs = geoNFTs?.result?.filter((geoNFT: any) => {
+    const matchesSearch = geoNFT.name?.toLowerCase().includes(searchTerm.toLowerCase()) ||
+                         geoNFT.description?.toLowerCase().includes(searchTerm.toLowerCase());
+    const matchesFilter = filterType === 'all' || geoNFT.type === filterType;
+    return matchesSearch && matchesFilter;
+  }) || [];
+
+  const sortedGeoNFTs = [...filteredGeoNFTs].sort((a: any, b: any) => {
+    switch (sortBy) {
+      case 'newest':
+        return new Date(b.createdOn || 0).getTime() - new Date(a.createdOn || 0).getTime();
+      case 'oldest':
+        return new Date(a.createdOn || 0).getTime() - new Date(b.createdOn || 0).getTime();
+      case 'name':
+        return (a.name || '').localeCompare(b.name || '');
+      case 'value':
+        return (b.value || 0) - (a.value || 0);
+      default:
+        return 0;
+    }
   });
 
-  const queryClient = useQueryClient();
-
-  const { data: geoNFTsData, isLoading, error, refetch } = useQuery(
-    'geoNFTs',
-    async () => {
-      try {
-        // Try to get real data first
-        const response = await starService.getAllGeoNFTs?.();
-        return response;
-      } catch (error) {
-        // Fallback to impressive demo data
-        console.log('Using demo GeoNFT data for investor presentation');
-        return {
-          result: [
-            {
-              id: '1',
-              name: 'Golden Gate Bridge',
-              description: 'Iconic suspension bridge spanning the Golden Gate strait',
-              imageUrl: 'https://images.unsplash.com/photo-1501594907352-04cda38ebc29?w=400&h=300&fit=crop',
-              latitude: 37.8199,
-              longitude: -122.4783,
-              location: 'San Francisco, CA',
-              category: 'Landmark',
-              rarity: 'Legendary',
-              price: 15.5,
-              owner: 'SF_Explorer',
-              createdAt: '2024-01-10',
-              isForSale: true,
-              views: 2840,
-              likes: 156,
-            },
-            {
-              id: '2',
-              name: 'Mount Fuji Peak',
-              description: 'Sacred mountain and active stratovolcano in Japan',
-              imageUrl: 'https://images.unsplash.com/photo-1490806843957-31f4c9a91c65?w=400&h=300&fit=crop',
-              latitude: 35.3606,
-              longitude: 138.7274,
-              location: 'Shizuoka, Japan',
-              category: 'Nature',
-              rarity: 'Epic',
-              price: 12.8,
-              owner: 'Mountain_Collector',
-              createdAt: '2024-01-12',
-              isForSale: false,
-              views: 1920,
-              likes: 98,
-            },
-            {
-              id: '3',
-              name: 'Times Square',
-              description: 'Major commercial intersection and tourist destination',
-              imageUrl: 'https://images.unsplash.com/photo-1496442226666-8d4d0e62e6e9?w=400&h=300&fit=crop',
-              latitude: 40.7580,
-              longitude: -73.9855,
-              location: 'New York, NY',
-              category: 'Urban',
-              rarity: 'Rare',
-              price: 8.2,
-              owner: 'NYC_Digital',
-              createdAt: '2024-01-15',
-              isForSale: true,
-              views: 1560,
-              likes: 87,
-            },
-            {
-              id: '4',
-              name: 'Machu Picchu',
-              description: '15th-century Inca citadel in the Andes Mountains',
-              imageUrl: 'https://images.unsplash.com/photo-1526392060635-9d6019884377?w=400&h=300&fit=crop',
-              latitude: -13.1631,
-              longitude: -72.5450,
-              location: 'Cusco, Peru',
-              category: 'Heritage',
-              rarity: 'Legendary',
-              price: 22.0,
-              owner: 'Ancient_Explorer',
-              createdAt: '2024-01-08',
-              isForSale: true,
-              views: 3200,
-              likes: 234,
-            },
-            {
-              id: '5',
-              name: 'Sahara Desert Dune',
-              description: 'Vast sand dune in the world\'s largest hot desert',
-              imageUrl: 'https://images.unsplash.com/photo-1509316975850-ff9c5deb0cd9?w=400&h=300&fit=crop',
-              latitude: 25.0000,
-              longitude: 0.0000,
-              location: 'Sahara Desert',
-              category: 'Nature',
-              rarity: 'Epic',
-              price: 9.5,
-              owner: 'Desert_Wanderer',
-              createdAt: '2024-01-14',
-              isForSale: false,
-              views: 980,
-              likes: 45,
-            },
-            {
-              id: '6',
-              name: 'Eiffel Tower',
-              description: 'Iron lattice tower on the Champ de Mars in Paris',
-              imageUrl: 'https://images.unsplash.com/photo-1511739001486-6bfe10ce785f?w=400&h=300&fit=crop',
-              latitude: 48.8584,
-              longitude: 2.2945,
-              location: 'Paris, France',
-              category: 'Landmark',
-              rarity: 'Legendary',
-              price: 18.7,
-              owner: 'Paris_Collector',
-              createdAt: '2024-01-11',
-              isForSale: true,
-              views: 4100,
-              likes: 289,
-            },
-          ]
-        };
-      }
-    },
-    {
-      refetchInterval: 30000,
-      refetchOnWindowFocus: true,
-    }
-  );
-
-  const createGeoNFTMutation = useMutation(
-    async (geoNFTData: Partial<GeoNFT>) => {
-      try {
-        return await starService.createGeoNFT?.(geoNFTData);
-      } catch (error) {
-        // For demo purposes, simulate success
-        toast.success('GeoNFT created successfully! (Demo Mode)');
-        return { success: true, id: Date.now().toString() };
-      }
-    },
-    {
-      onSuccess: () => {
-        queryClient.invalidateQueries('geoNFTs');
-        setCreateDialogOpen(false);
-        setNewGeoNFT({
-          name: '',
-          description: '',
-          imageUrl: '',
-          latitude: 0,
-          longitude: 0,
-          location: '',
-          category: 'Landmark',
-          rarity: 'Common',
-          price: 0,
-        });
-      },
-      onError: () => {
-        toast.error('Failed to create GeoNFT');
-      },
-    }
-  );
-
-  const deleteGeoNFTMutation = useMutation(
-    async (id: string) => {
-      try {
-        return await starService.deleteGeoNFT?.(id);
-      } catch (error) {
-        // For demo purposes, simulate success
-        toast.success('GeoNFT deleted successfully! (Demo Mode)');
-        return { success: true };
-      }
-    },
-    {
-      onSuccess: () => {
-        queryClient.invalidateQueries('geoNFTs');
-      },
-      onError: () => {
-        toast.error('Failed to delete GeoNFT');
-      },
-    }
-  );
-
-  const handleCreateGeoNFT = () => {
-    if (!newGeoNFT.name || !newGeoNFT.description || !newGeoNFT.location) {
-      toast.error('Please fill in all required fields');
-      return;
-    }
-    createGeoNFTMutation.mutate(newGeoNFT);
-  };
-
-  const handleDeleteGeoNFT = (id: string) => {
-    if (window.confirm('Are you sure you want to delete this GeoNFT?')) {
-      deleteGeoNFTMutation.mutate(id);
-    }
-  };
-
-  const getRarityColor = (rarity: string) => {
-    switch (rarity.toLowerCase()) {
-      case 'legendary': return '#9c27b0';
-      case 'epic': return '#ff9800';
-      case 'rare': return '#2196f3';
-      case 'uncommon': return '#4caf50';
-      case 'common': return '#9e9e9e';
-      default: return '#9e9e9e';
-    }
-  };
-
-  const getCategoryIcon = (category: string) => {
-    switch (category.toLowerCase()) {
-      case 'landmark': return <Public />;
-      case 'nature': return <Nature />;
-      case 'urban': return <Business />;
-      case 'heritage': return <Terrain />;
-      default: return <LocationOn />;
-    }
-  };
-
-  const categories = ['all', 'Landmark', 'Nature', 'Urban', 'Heritage'];
-  
-  // Filter by view scope first, then by category
-  const getFilteredGeoNFTs = () => {
-    let filtered = geoNFTsData?.result || [];
-    
-    // Apply view scope filter
-    if (viewScope === 'installed') {
-      filtered = filtered.filter((geoNFT: any) => geoNFT.isInstalled);
-    } else if (viewScope === 'mine') {
-      // For demo, show GeoNFTs created by current user
-      filtered = filtered.filter((geoNFT: any) => geoNFT.owner === 'SF_Explorer');
-    }
-    
-    // Apply category filter
-    if (filterCategory !== 'all') {
-      filtered = filtered.filter((geoNFT: any) => geoNFT.category === filterCategory);
-    }
-    
-    return filtered;
-  };
-  
-  const filteredGeoNFTs = getFilteredGeoNFTs();
-
-  const containerVariants = {
-    hidden: { opacity: 0 },
-    visible: {
-      opacity: 1,
-      transition: {
-        duration: 0.5,
-        staggerChildren: 0.1,
-      },
-    },
-  };
-
-  const itemVariants = {
-    hidden: { opacity: 0, y: 20 },
-    visible: { opacity: 1, y: 0 },
+  const geoNFTStats = {
+    total: geoNFTs?.result?.length || 0,
+    published: geoNFTs?.result?.filter((geoNFT: any) => geoNFT.isPublished).length || 0,
+    totalValue: geoNFTs?.result?.reduce((sum: number, geoNFT: any) => sum + (geoNFT.value || 0), 0) || 0,
+    averageRating: geoNFTs?.result?.reduce((sum: number, geoNFT: any) => sum + (geoNFT.rating || 0), 0) / (geoNFTs?.result?.length || 1) || 0,
   };
 
   return (
-    <motion.div
-      variants={containerVariants}
-      initial="hidden"
-      animate="visible"
-    >
-      <>
-      <Box sx={{ mb: 4, mt: 4 }}>
-        <Box sx={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', mb: 3 }}>
-          <Box>
-            <Typography variant="h4" gutterBottom className="page-heading">
-              GeoNFTs
-            </Typography>
-            <Typography variant="subtitle1" color="text.secondary">
-              Location-based NFTs and geo-spatial assets
+    <Box sx={{ flexGrow: 1, p: 3 }}>
+      {/* Header */}
+      <Box sx={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', mb: 3 }}>
+        <Box>
+          <Typography variant="h4" component="h1" gutterBottom>
+            GeoNFTs
+          </Typography>
+          <Typography variant="body1" color="text.secondary">
+            Manage your Location-based Non-Fungible Tokens
+          </Typography>
+          <Box sx={{ mt: 1, p: 2, bgcolor: '#0d47a1', color: 'white', borderRadius: 1, display: 'flex', alignItems: 'center', gap: 1 }}>
+            <Info sx={{ color: 'white' }} />
+            <Typography variant="body2" sx={{ color: 'white' }}>
+              Create, map and manage GeoNFTs. Data updates in real-time.
             </Typography>
           </Box>
-          <Box sx={{ display: 'flex', gap: 2 }}>
-            <Button
-              variant="outlined"
-              startIcon={<Refresh />}
-              onClick={() => refetch()}
-              disabled={isLoading}
-            >
-              Refresh
-            </Button>
-            <Button
-              variant="contained"
-              startIcon={<Add />}
-              onClick={() => setCreateDialogOpen(true)}
-              sx={{
-                bgcolor: '#1976d2',
-                '&:hover': {
-                  bgcolor: '#1565c0',
-                },
-              }}
-            >
-              Create GeoNFT
-            </Button>
-          </Box>
         </Box>
-
-        {/* Filter */}
-        <Box sx={{ mb: 3, display: 'flex', alignItems: 'center', gap: 2 }}>
-          <FormControl size="small" sx={{ minWidth: 140 }}>
-            <InputLabel>View</InputLabel>
-            <Select
-              value={viewScope}
-              label="View"
-              onChange={(e) => setViewScope(e.target.value)}
-            >
-              <MenuItem value="all">All</MenuItem>
-              <MenuItem value="installed">Installed</MenuItem>
-              <MenuItem value="mine">My GeoNFTs</MenuItem>
-            </Select>
-          </FormControl>
-          
-          <FilterList color="action" />
-          <FormControl size="small" sx={{ minWidth: 150 }}>
-            <InputLabel>Category</InputLabel>
-            <Select
-              value={filterCategory}
-              label="Category"
-              onChange={(e) => setFilterCategory(e.target.value)}
-            >
-              {categories.map((category) => (
-                <MenuItem key={category} value={category}>
-                  {category === 'all' ? 'All Categories' : category}
-                </MenuItem>
-              ))}
-            </Select>
-          </FormControl>
-        </Box>
+        <Button
+          variant="contained"
+          startIcon={<Add />}
+          onClick={() => setCreateDialogOpen(true)}
+          sx={{ borderRadius: 2 }}
+        >
+          Create GeoNFT
+        </Button>
       </Box>
 
-      {error && (
-        <Alert severity="error" sx={{ mb: 3 }}>
-          Failed to load GeoNFTs: {error instanceof Error ? error.message : 'Unknown error'}
-        </Alert>
-      )}
+      {/* Stats Cards */}
+      <Grid container spacing={3} sx={{ mb: 3 }}>
+        <Grid item xs={12} sm={6} md={3}>
+          <Card>
+            <CardContent>
+              <Box sx={{ display: 'flex', alignItems: 'center' }}>
+                <LocationOn color="primary" sx={{ mr: 2 }} />
+                <Box>
+                  <Typography variant="h6">{geoNFTStats.total}</Typography>
+                  <Typography variant="body2" color="text.secondary">
+                    Total GeoNFTs
+                  </Typography>
+                </Box>
+              </Box>
+            </CardContent>
+          </Card>
+        </Grid>
+        <Grid item xs={12} sm={6} md={3}>
+          <Card>
+            <CardContent>
+              <Box sx={{ display: 'flex', alignItems: 'center' }}>
+                <Upload color="success" sx={{ mr: 2 }} />
+                <Box>
+                  <Typography variant="h6">{geoNFTStats.published}</Typography>
+                  <Typography variant="body2" color="text.secondary">
+                    Published
+                  </Typography>
+                </Box>
+              </Box>
+            </CardContent>
+          </Card>
+        </Grid>
+        <Grid item xs={12} sm={6} md={3}>
+          <Card>
+            <CardContent>
+              <Box sx={{ display: 'flex', alignItems: 'center' }}>
+                <MonetizationOn color="warning" sx={{ mr: 2 }} />
+                <Box>
+                  <Typography variant="h6">{geoNFTStats.totalValue.toFixed(2)} ETH</Typography>
+                  <Typography variant="body2" color="text.secondary">
+                    Total Value
+                  </Typography>
+                </Box>
+              </Box>
+            </CardContent>
+          </Card>
+        </Grid>
+        <Grid item xs={12} sm={6} md={3}>
+          <Card>
+            <CardContent>
+              <Box sx={{ display: 'flex', alignItems: 'center' }}>
+                <Star color="info" sx={{ mr: 2 }} />
+                <Box>
+                  <Typography variant="h6">{geoNFTStats.averageRating.toFixed(1)}</Typography>
+                  <Typography variant="body2" color="text.secondary">
+                    Avg Rating
+                  </Typography>
+                </Box>
+              </Box>
+            </CardContent>
+          </Card>
+        </Grid>
+      </Grid>
 
-      {isLoading ? (
-        <Box sx={{ display: 'flex', justifyContent: 'center', py: 4 }}>
-          <CircularProgress />
-        </Box>
-      ) : (
-        <Grid container spacing={3}>
-          {filteredGeoNFTs.map((geoNFT: any, index: number) => (
-            <Grid item xs={12} sm={6} md={4} lg={3} key={geoNFT.id}>
-              <motion.div
-                variants={itemVariants}
-                whileHover={{ 
-                  scale: 1.05,
-                  transition: { duration: 0.2 }
+      {/* Filters and Search */}
+      <Card sx={{ mb: 3 }}>
+        <CardContent>
+          <Grid container spacing={2} alignItems="center">
+            <Grid item xs={12} md={4}>
+              <TextField
+                fullWidth
+                placeholder="Search GeoNFTs..."
+                value={searchTerm}
+                onChange={(e) => setSearchTerm(e.target.value)}
+                InputProps={{
+                  startAdornment: <Search sx={{ mr: 1, color: 'text.secondary' }} />,
                 }}
-                whileTap={{ scale: 0.95 }}
-              >
-                <Card 
-                  sx={{ height: '100%', position: 'relative', cursor: 'pointer' }}
-                  onClick={() => navigate(`/geonfts/${geoNFT.id}`)}
+              />
+            </Grid>
+            <Grid item xs={12} md={3}>
+              <FormControl fullWidth>
+                <InputLabel>Filter by Type</InputLabel>
+                <Select
+                  value={filterType}
+                  onChange={(e) => setFilterType(e.target.value)}
                 >
-                  <Box sx={{ position: 'relative' }}>
-                    <img
-                      src={geoNFT.imageUrl}
-                      alt={geoNFT.name}
-                      style={{
-                        width: '100%',
-                        height: 200,
-                        objectFit: 'cover',
-                        borderTopLeftRadius: 4,
-                        borderTopRightRadius: 4,
-                      }}
-                    />
-                    <Chip
-                      label={geoNFT.rarity}
-                      size="small"
-                      sx={{
-                        position: 'absolute',
-                        top: 8,
-                        right: 8,
-                        bgcolor: getRarityColor(geoNFT.rarity),
-                        color: 'white',
-                        fontWeight: 'bold',
-                      }}
-                    />
-                    {geoNFT.isForSale && (
-                      <Chip
-                        label="FOR SALE"
-                        size="small"
-                        color="success"
-                        sx={{
-                          position: 'absolute',
-                          top: 8,
-                          left: 8,
-                          fontWeight: 'bold',
-                        }}
-                      />
-                    )}
-                    {geoNFT.isInstalled && (
-                      <Chip
-                        label="Installed"
-                        size="small"
-                        color="success"
-                        sx={{
-                          position: 'absolute',
-                          bottom: 8,
-                          right: 8,
-                          fontWeight: 'bold',
-                        }}
-                      />
-                    )}
-                  </Box>
-                  
-                  <CardContent>
-                    <Box sx={{ display: 'flex', alignItems: 'center', mb: 1 }}>
-                      {getCategoryIcon(geoNFT.category)}
-                      <Typography variant="h6" sx={{ ml: 1, flexGrow: 1 }}>
-                        {geoNFT.name}
+                  <MenuItem value="all">All Types</MenuItem>
+                  <MenuItem value="Landmark">Landmark</MenuItem>
+                  <MenuItem value="Event">Event</MenuItem>
+                  <MenuItem value="Art">Art</MenuItem>
+                  <MenuItem value="Experience">Experience</MenuItem>
+                </Select>
+              </FormControl>
+            </Grid>
+            <Grid item xs={12} md={3}>
+              <FormControl fullWidth>
+                <InputLabel>Sort by</InputLabel>
+                <Select
+                  value={sortBy}
+                  onChange={(e) => setSortBy(e.target.value)}
+                >
+                  <MenuItem value="newest">Newest</MenuItem>
+                  <MenuItem value="oldest">Oldest</MenuItem>
+                  <MenuItem value="name">Name</MenuItem>
+                  <MenuItem value="value">Value</MenuItem>
+                </Select>
+              </FormControl>
+            </Grid>
+            <Grid item xs={12} md={2}>
+              <Button
+                fullWidth
+                variant="outlined"
+                startIcon={<Map />}
+              >
+                View Map
+              </Button>
+            </Grid>
+          </Grid>
+        </CardContent>
+      </Card>
+
+      {/* GeoNFTs Grid */}
+      <Grid container spacing={3}>
+        {isLoading ? (
+          Array.from({ length: 6 }).map((_, index) => (
+            <Grid item xs={12} sm={6} md={4} key={index}>
+              <Card>
+                <CardContent>
+                  <Box sx={{ display: 'flex', alignItems: 'center', mb: 2 }}>
+                    <Avatar sx={{ mr: 2 }} />
+                    <Box sx={{ flexGrow: 1 }}>
+                      <Typography variant="h6" sx={{ mb: 1 }}>
+                        Loading...
+                      </Typography>
+                      <Typography variant="body2" color="text.secondary">
+                        Loading GeoNFT details...
                       </Typography>
                     </Box>
-                    
+                  </Box>
+                </CardContent>
+              </Card>
+            </Grid>
+          ))
+        ) : error ? (
+          <Grid item xs={12}>
+            <Card>
+              <CardContent sx={{ textAlign: 'center', py: 4 }}>
+                <Typography variant="h6" color="error" gutterBottom>
+                  Failed to load GeoNFTs
+                </Typography>
+                <Typography variant="body2" color="text.secondary">
+                  {error instanceof Error ? error.message : 'An error occurred'}
+                </Typography>
+              </CardContent>
+            </Card>
+          </Grid>
+        ) : sortedGeoNFTs.length === 0 ? (
+          <Grid item xs={12}>
+            <Card>
+              <CardContent sx={{ textAlign: 'center', py: 4 }}>
+                <LocationOn sx={{ fontSize: 64, color: 'text.secondary', mb: 2 }} />
+                <Typography variant="h6" gutterBottom>
+                  No GeoNFTs found
+                </Typography>
+                <Typography variant="body2" color="text.secondary" sx={{ mb: 3 }}>
+                  {searchTerm ? 'Try adjusting your search criteria' : 'Create your first GeoNFT to get started'}
+                </Typography>
+                <Button
+                  variant="contained"
+                  startIcon={<Add />}
+                  onClick={() => setCreateDialogOpen(true)}
+                >
+                  Create GeoNFT
+                </Button>
+              </CardContent>
+            </Card>
+          </Grid>
+        ) : (
+          sortedGeoNFTs.map((geoNFT: any) => (
+            <Grid item xs={12} sm={6} md={4} key={geoNFT.id}>
+              <motion.div
+                initial={{ opacity: 0, y: 20 }}
+                animate={{ opacity: 1, y: 0 }}
+                transition={{ duration: 0.3 }}
+              >
+                <Card sx={{ height: '100%', display: 'flex', flexDirection: 'column' }}>
+                  <CardMedia
+                    component="img"
+                    height="200"
+                    image={geoNFT.imageUrl || '/api/placeholder/400/200'}
+                    alt={geoNFT.name}
+                    sx={{ objectFit: 'cover' }}
+                  />
+                  <CardContent sx={{ flexGrow: 1 }}>
+                    <Box sx={{ display: 'flex', justifyContent: 'space-between', alignItems: 'flex-start', mb: 2 }}>
+                      <Typography variant="h6" component="h3" noWrap>
+                        {geoNFT.name}
+                      </Typography>
+                      <IconButton
+                        size="small"
+                        onClick={(e) => handleMenuClick(e, geoNFT)}
+                      >
+                        <MoreVert />
+                      </IconButton>
+                    </Box>
                     <Typography variant="body2" color="text.secondary" sx={{ mb: 2 }}>
                       {geoNFT.description}
                     </Typography>
-                    
-                    <Box sx={{ display: 'flex', alignItems: 'center', mb: 1 }}>
-                      <LocationOn sx={{ fontSize: 16, mr: 1, color: 'text.secondary' }} />
-                      <Typography variant="body2" color="text.secondary">
-                        {geoNFT.location}
-                      </Typography>
-                    </Box>
-                    
                     <Box sx={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', mb: 2 }}>
-                      <Typography variant="h6" color="primary" fontWeight="bold">
-                        {geoNFT.price} ETH
-                      </Typography>
-                      <Box sx={{ display: 'flex', gap: 2, alignItems: 'center' }}>
-                        <Box sx={{ display: 'flex', alignItems: 'center', gap: 0.5 }}>
-                          <Visibility sx={{ fontSize: 16, color: 'text.secondary' }} />
-                          <Typography variant="body2" color="text.secondary">
-                            {geoNFT.views}
-                          </Typography>
-                        </Box>
-                        <Box sx={{ display: 'flex', alignItems: 'center', gap: 0.5 }}>
-                          <Typography variant="body2">❤️</Typography>
-                          <Typography variant="body2" color="text.secondary">
-                            {geoNFT.likes}
-                          </Typography>
-                        </Box>
+                      <Chip
+                        label={geoNFT.type || 'GeoNFT'}
+                        size="small"
+                        color="primary"
+                        variant="outlined"
+                      />
+                      <Box sx={{ display: 'flex', alignItems: 'center' }}>
+                        <Star sx={{ fontSize: 16, mr: 0.5, color: 'warning.main' }} />
+                        <Typography variant="body2">
+                          {geoNFT.rating || 0}
+                        </Typography>
                       </Box>
                     </Box>
-                    
+                    <Box sx={{ display: 'flex', alignItems: 'center', mb: 2 }}>
+                      <LocationOn sx={{ fontSize: 16, mr: 0.5, color: 'text.secondary' }} />
+                      <Typography variant="body2" color="text.secondary">
+                        {geoNFT.latitude?.toFixed(4)}, {geoNFT.longitude?.toFixed(4)}
+                      </Typography>
+                    </Box>
                     <Box sx={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}>
-                      <Button
-                        variant="outlined"
-                        size="small"
-                        startIcon={<Visibility />}
-                        onClick={() => toast.success('Viewing GeoNFT details')}
-                      >
-                        View
-                      </Button>
-                      <IconButton
-                        size="small"
-                        onClick={() => handleDeleteGeoNFT(geoNFT.id)}
-                        disabled={deleteGeoNFTMutation.isLoading}
-                        color="error"
-                      >
-                        <Delete />
-                      </IconButton>
+                      <Typography variant="h6" color="primary">
+                        {geoNFT.value || '0'} ETH
+                      </Typography>
+                      <Box sx={{ display: 'flex', alignItems: 'center' }}>
+                        <Typography variant="body2" color="text.secondary" sx={{ mr: 1 }}>
+                          {geoNFT.downloads || 0} downloads
+                        </Typography>
+                      </Box>
                     </Box>
                   </CardContent>
+                  <Divider />
+                  <CardActions sx={{ justifyContent: 'space-between', p: 2 }}>
+                    <Button
+                      size="small"
+                      startIcon={<Visibility />}
+                      onClick={() => navigate(`/geonfts/${geoNFT.id}`)}
+                    >
+                      View
+                    </Button>
+                    <Box>
+                      <IconButton
+                        size="small"
+                        onClick={() => handleDownloadGeoNFT(geoNFT.id)}
+                        disabled={downloadGeoNFTMutation.isLoading}
+                      >
+                        <Download />
+                      </IconButton>
+                      <IconButton
+                        size="small"
+                        onClick={() => handlePublishGeoNFT(geoNFT.id)}
+                        disabled={publishGeoNFTMutation.isLoading}
+                      >
+                        <Upload />
+                      </IconButton>
+                    </Box>
+                  </CardActions>
                 </Card>
               </motion.div>
             </Grid>
-          ))}
-        </Grid>
-      )}
+          ))
+        )}
+      </Grid>
 
       {/* Create GeoNFT Dialog */}
-      <Dialog open={createDialogOpen} onClose={() => setCreateDialogOpen(false)} maxWidth="md" fullWidth>
+      <Dialog
+        open={createDialogOpen}
+        onClose={() => setCreateDialogOpen(false)}
+        maxWidth="md"
+        fullWidth
+      >
         <DialogTitle>Create New GeoNFT</DialogTitle>
         <DialogContent>
-          <Box sx={{ display: 'flex', flexDirection: 'column', gap: 2, pt: 1 }}>
-            <TextField
-              label="Name"
-              value={newGeoNFT.name}
-              onChange={(e) => setNewGeoNFT({ ...newGeoNFT, name: e.target.value })}
-              fullWidth
-              required
-            />
-            <TextField
-              label="Description"
-              value={newGeoNFT.description}
-              onChange={(e) => setNewGeoNFT({ ...newGeoNFT, description: e.target.value })}
-              fullWidth
-              multiline
-              rows={3}
-              required
-            />
-            <TextField
-              label="Image URL"
-              value={newGeoNFT.imageUrl}
-              onChange={(e) => setNewGeoNFT({ ...newGeoNFT, imageUrl: e.target.value })}
-              fullWidth
-            />
-            <Box sx={{ display: 'flex', gap: 2 }}>
+          <Grid container spacing={2} sx={{ mt: 1 }}>
+            <Grid item xs={12}>
               <TextField
+                fullWidth
+                label="GeoNFT Name"
+                placeholder="Enter GeoNFT name"
+                required
+              />
+            </Grid>
+            <Grid item xs={12}>
+              <TextField
+                fullWidth
+                label="Description"
+                placeholder="Enter GeoNFT description"
+                multiline
+                rows={3}
+              />
+            </Grid>
+            <Grid item xs={12} sm={6}>
+              <FormControl fullWidth>
+                <InputLabel>Type</InputLabel>
+                <Select defaultValue="Landmark">
+                  <MenuItem value="Landmark">Landmark</MenuItem>
+                  <MenuItem value="Event">Event</MenuItem>
+                  <MenuItem value="Art">Art</MenuItem>
+                  <MenuItem value="Experience">Experience</MenuItem>
+                </Select>
+              </FormControl>
+            </Grid>
+            <Grid item xs={12} sm={6}>
+              <TextField
+                fullWidth
+                label="Value (ETH)"
+                type="number"
+                placeholder="0.1"
+              />
+            </Grid>
+            <Grid item xs={12} sm={6}>
+              <TextField
+                fullWidth
                 label="Latitude"
                 type="number"
-                value={newGeoNFT.latitude}
-                onChange={(e) => setNewGeoNFT({ ...newGeoNFT, latitude: parseFloat(e.target.value) })}
-                fullWidth
+                placeholder="40.7128"
               />
+            </Grid>
+            <Grid item xs={12} sm={6}>
               <TextField
+                fullWidth
                 label="Longitude"
                 type="number"
-                value={newGeoNFT.longitude}
-                onChange={(e) => setNewGeoNFT({ ...newGeoNFT, longitude: parseFloat(e.target.value) })}
-                fullWidth
+                placeholder="-74.0060"
               />
-            </Box>
-            <TextField
-              label="Location"
-              value={newGeoNFT.location}
-              onChange={(e) => setNewGeoNFT({ ...newGeoNFT, location: e.target.value })}
-              fullWidth
-              required
-            />
-            <Box sx={{ display: 'flex', gap: 2 }}>
-              <FormControl fullWidth>
-                <InputLabel>Category</InputLabel>
-                <Select
-                  value={newGeoNFT.category}
-                  label="Category"
-                  onChange={(e) => setNewGeoNFT({ ...newGeoNFT, category: e.target.value })}
-                >
-                  <MenuItem value="Landmark">Landmark</MenuItem>
-                  <MenuItem value="Nature">Nature</MenuItem>
-                  <MenuItem value="Urban">Urban</MenuItem>
-                  <MenuItem value="Heritage">Heritage</MenuItem>
-                </Select>
-              </FormControl>
-              <FormControl fullWidth>
-                <InputLabel>Rarity</InputLabel>
-                <Select
-                  value={newGeoNFT.rarity}
-                  label="Rarity"
-                  onChange={(e) => setNewGeoNFT({ ...newGeoNFT, rarity: e.target.value })}
-                >
-                  <MenuItem value="Common">Common</MenuItem>
-                  <MenuItem value="Uncommon">Uncommon</MenuItem>
-                  <MenuItem value="Rare">Rare</MenuItem>
-                  <MenuItem value="Epic">Epic</MenuItem>
-                  <MenuItem value="Legendary">Legendary</MenuItem>
-                </Select>
-              </FormControl>
-            </Box>
-            <TextField
-              label="Price (ETH)"
-              type="number"
-              value={newGeoNFT.price}
-              onChange={(e) => setNewGeoNFT({ ...newGeoNFT, price: parseFloat(e.target.value) })}
-              fullWidth
-            />
-          </Box>
+            </Grid>
+            <Grid item xs={12}>
+              <TextField
+                fullWidth
+                label="Image URL"
+                placeholder="https://example.com/image.png"
+              />
+            </Grid>
+          </Grid>
         </DialogContent>
         <DialogActions>
-          <Button onClick={() => setCreateDialogOpen(false)}>Cancel</Button>
+          <Button onClick={() => setCreateDialogOpen(false)}>
+            Cancel
+          </Button>
           <Button
-            onClick={handleCreateGeoNFT}
             variant="contained"
+            onClick={() => handleCreateGeoNFT({})}
             disabled={createGeoNFTMutation.isLoading}
           >
             {createGeoNFTMutation.isLoading ? 'Creating...' : 'Create GeoNFT'}
           </Button>
         </DialogActions>
       </Dialog>
-      </>
-    </motion.div>
+
+      {/* Context Menu */}
+      <Menu
+        anchorEl={anchorEl}
+        open={Boolean(anchorEl)}
+        onClose={handleMenuClose}
+      >
+        <MenuItem onClick={() => {
+          if (selectedGeoNFT) navigate(`/geonfts/${selectedGeoNFT.id}`);
+          handleMenuClose();
+        }}>
+          <Visibility sx={{ mr: 1 }} />
+          View Details
+        </MenuItem>
+        <MenuItem onClick={() => {
+          if (selectedGeoNFT) handleDownloadGeoNFT(selectedGeoNFT.id);
+          handleMenuClose();
+        }}>
+          <Download sx={{ mr: 1 }} />
+          Download
+        </MenuItem>
+        <MenuItem onClick={() => {
+          if (selectedGeoNFT) handlePublishGeoNFT(selectedGeoNFT.id);
+          handleMenuClose();
+        }}>
+          <Upload sx={{ mr: 1 }} />
+          Publish
+        </MenuItem>
+        <MenuItem onClick={handleMenuClose}>
+          <Edit sx={{ mr: 1 }} />
+          Edit
+        </MenuItem>
+        <MenuItem onClick={handleMenuClose} sx={{ color: 'error.main' }}>
+          <Delete sx={{ mr: 1 }} />
+          Delete
+        </MenuItem>
+      </Menu>
+
+      {/* Floating Action Button */}
+      <Fab
+        color="primary"
+        aria-label="create geonft"
+        sx={{ position: 'fixed', bottom: 16, right: 16 }}
+        onClick={() => setCreateDialogOpen(true)}
+      >
+        <Add />
+      </Fab>
+    </Box>
   );
 };
 
