@@ -30,6 +30,31 @@ import {
   AccountCircle,
   Login,
   PersonAdd,
+  AutoAwesome,
+  Memory,
+  Storage,
+  Psychology,
+  Build,
+  Apps,
+  Image,
+  LocationOn,
+  Assignment,
+  FlightTakeoff,
+  MenuBook,
+  Inventory,
+  SpaceDashboard,
+  DataObject,
+  LibraryBooks,
+  Extension,
+  AccountBalanceWallet,
+  Terminal,
+  Gamepad,
+  ViewInAr,
+  School,
+  Business,
+  ShoppingCart,
+  LocalHospital,
+  AttachMoney,
 } from '@mui/icons-material';
 import { motion } from 'framer-motion';
 import { useNavigate } from 'react-router-dom';
@@ -39,42 +64,87 @@ const HomePage: React.FC = () => {
 
   const features = [
     {
-      icon: <Star sx={{ fontSize: 40, color: '#ffd700' }} />,
-      title: 'OASIS Integration',
-      description: 'Seamlessly connect with the OASIS universe and access all its features',
+      icon: <AutoAwesome sx={{ fontSize: 40, color: '#ffd700' }} />,
+      title: 'OASIS HyperDrive',
+      description: 'Revolutionary auto-failover system that intelligently switches between Web2 and Web3 providers for optimal performance',
     },
     {
       icon: <Rocket sx={{ fontSize: 40, color: '#ff6b35' }} />,
-      title: 'Lightning Fast',
-      description: 'Built for speed with cutting-edge technology and optimized performance',
+      title: 'Universal Data Aggregation',
+      description: 'Single API that connects everything to everything - the first universal Web2/Web3 aggregation system',
     },
     {
       icon: <Security sx={{ fontSize: 40, color: '#4caf50' }} />,
-      title: 'Secure & Private',
-      description: 'Enterprise-grade security with end-to-end encryption and privacy protection',
+      title: 'Zero-Downtime Architecture',
+      description: 'Impossible to shutdown with distributed, redundant architecture and hot-swappable providers',
     },
     {
-      icon: <Public sx={{ fontSize: 40, color: '#2196f3' }} />,
-      title: 'Global Network',
-      description: 'Connect with users worldwide in the largest virtual reality network',
+      icon: <Build sx={{ fontSize: 40, color: '#2196f3' }} />,
+      title: 'STAR ODK Platform',
+      description: 'No/Low-code metaverse development with OAPPs, missions, NFTs, and comprehensive gamification',
     },
     {
-      icon: <Cloud sx={{ fontSize: 40, color: '#9c27b0' }} />,
-      title: 'Cloud Powered',
-      description: 'Leverage the power of cloud computing for unlimited scalability',
+      icon: <Memory sx={{ fontSize: 40, color: '#9c27b0' }} />,
+      title: 'Future-Proof Technology',
+      description: 'Never learn new tech stacks again - universal API abstraction with HOT swappable plugins',
     },
     {
-      icon: <Science sx={{ fontSize: 40, color: '#ff9800' }} />,
-      title: 'AI Enhanced',
-      description: 'Advanced AI integration for intelligent automation and assistance',
+      icon: <Psychology sx={{ fontSize: 40, color: '#ff9800' }} />,
+      title: 'Karma & Reputation System',
+      description: 'Universal reputation tracking across all platforms with cross-platform contribution rewards',
+    },
+    {
+      icon: <Apps sx={{ fontSize: 40, color: '#e91e63' }} />,
+      title: 'STARNET Holons Linking',
+        description: 'Next Generation linking system - any STARNET Holon can be connected to any other STARNET Holon as dependencies creating unique & infinite possibilities!',
+    },
+    {
+      icon: <Code sx={{ fontSize: 40, color: '#795548' }} />,
+      title: 'STAR CLI',
+      description: 'Revolutionary Interoperable Low/No Code Generator for all metaverses, games, apps, sites, and platforms',
+    },
+    {
+      icon: <AccountBalanceWallet sx={{ fontSize: 40, color: '#00bcd4' }} />,
+      title: 'OASIS Universal Wallet',
+      description: 'Unified digital asset management across 50+ blockchain networks with cross-chain support and DeFi integration',
+    },
+    {
+      icon: <Gamepad sx={{ fontSize: 40, color: '#4caf50' }} />,
+      title: 'Our World AR Game',
+      description: 'Groundbreaking AR geo-location game that encourages environmental stewardship and community service',
+    },
+    {
+      icon: <ViewInAr sx={{ fontSize: 40, color: '#9c27b0' }} />,
+      title: 'One World MMORPG',
+      description: 'Benevolent MMORPG with optional VR, similar to Minecraft and Pax Dei with infinite building possibilities',
+    },
+    {
+      icon: <Public sx={{ fontSize: 40, color: '#ff5722' }} />,
+      title: 'Cross-Platform Universal System',
+      description: 'ALL STARNET Holons can be shared across ANY OAPP (apps, games, sites, services, platforms) with infinite use cases',
+    },
+    {
+      icon: <Terminal sx={{ fontSize: 40, color: '#795548' }} />,
+      title: 'HoloNET Integration',
+      description: 'World-first .NET and Unity client for Holochain, bringing P2P architecture to mainstream development',
+    },
+    {
+      icon: <Image sx={{ fontSize: 40, color: '#ff9800' }} />,
+      title: 'Revolutionary NFT System',
+      description: 'Cross-chain NFTs with shared metadata, Geo-NFTs, and universal NFT standard across all platforms',
+    },
+    {
+      icon: <Storage sx={{ fontSize: 40, color: '#607d8b' }} />,
+      title: 'OASIS COSMIC ORM',
+      description: 'Universal data abstraction layer for all Web2/Web3 providers with simple .Load(), .Save(), .Delete() commands',
     },
   ];
 
   const stats = [
-    { label: 'Active Users', value: '2.5M+', color: '#00bcd4' },
-    { label: 'Virtual Worlds', value: '10K+', color: '#00acc1' },
-    { label: 'Transactions', value: '$50M+', color: '#0097a7' },
-    { label: 'Countries', value: '180+', color: '#00838f' },
+    { label: 'OASIS Users', value: '2.5M+', color: '#00bcd4' },
+    { label: 'Supported Providers', value: '50+', color: '#00acc1' },
+    { label: 'OAPPs Deployed', value: '15K+', color: '#0097a7' },
+    { label: 'Karma Points', value: '12.5M+', color: '#00838f' },
   ];
 
   const containerVariants = {
@@ -141,7 +211,7 @@ const HomePage: React.FC = () => {
                   whileHover={{ scale: 1.05 }}
                 >
                   <img 
-                    src="/star-logo.png" 
+                    src="/star-logo-1.png" 
                     alt="STAR Logo" 
                     style={{ 
                       maxWidth: '300px', 
@@ -153,12 +223,13 @@ const HomePage: React.FC = () => {
               </Box>
               
               <Typography variant="h2" component="h1" gutterBottom sx={{ fontWeight: 'bold', mb: 3 }}>
-                Welcome to STARNET
+                Welcome to The OASIS & STARNET!
               </Typography>
               
-              <Typography variant="h5" sx={{ mb: 4, opacity: 0.9, maxWidth: 600, mx: 'auto' }}>
-                The ultimate platform for the OASIS Omniverse. Connect, create, and explore 
-                in the most advanced interoperable AR/VR/IR Metaverse ecosystem ever built.
+              <Typography variant="h5" sx={{ mb: 4, opacity: 0.9, maxWidth: 800, mx: 'auto' }}>
+                The Universal Web4/Web5 Infrastructure that unifies all Web2 and Web3 technologies 
+                into a single, intelligent, auto-failover system. The first universal API that 
+                connects everything to everything.
               </Typography>
               
               <Stack direction="row" spacing={2} justifyContent="center" sx={{ mb: 6 }}>
@@ -206,7 +277,7 @@ const HomePage: React.FC = () => {
         <Container maxWidth="lg">
           <motion.div variants={itemVariants}>
             <Typography variant="h4" align="center" gutterBottom sx={{ mb: 6, fontWeight: 'bold' }}>
-              Trusted by Millions Worldwide
+              OASIS Ecosystem Metrics
             </Typography>
             
             <Grid container spacing={4}>
@@ -239,7 +310,7 @@ const HomePage: React.FC = () => {
         <Container maxWidth="lg">
           <motion.div variants={itemVariants}>
             <Typography variant="h4" align="center" gutterBottom sx={{ mb: 8, fontWeight: 'bold' }}>
-              Why Choose STARNET?
+              Revolutionary OASIS Technology
             </Typography>
             
             <Grid container spacing={4}>
@@ -271,12 +342,157 @@ const HomePage: React.FC = () => {
         </Container>
       </Box>
 
-      {/* Platform Overview */}
+      {/* Infinite Use Cases Section */}
       <Box sx={{ py: 12, bgcolor: 'background.default' }}>
         <Container maxWidth="lg">
           <motion.div variants={itemVariants}>
             <Typography variant="h4" align="center" gutterBottom sx={{ mb: 8, fontWeight: 'bold' }}>
-              Complete OASIS Platform
+              Infinite Use Cases - The Future of Everything
+            </Typography>
+            
+            <Typography variant="h6" align="center" color="text.secondary" sx={{ mb: 6 }}>
+              ALL STARNET Holons can be shared across ANY OAPP (apps, games, sites, services, platforms, etc.) - 
+              creating infinite possibilities for games, businesses, shops, e-commerce, finance, education, healthcare, and everything else!
+            </Typography>
+            
+            <Grid container spacing={4}>
+              {[
+                { name: 'Games', description: 'AR/VR Games, MMORPGs, Mobile Games', icon: <Gamepad />, color: '#4caf50' },
+                { name: 'Business', description: 'Enterprise Applications, CRM, ERP', icon: <Business />, color: '#2196f3' },
+                { name: 'E-Commerce', description: 'Online Stores, Marketplaces, Shopping', icon: <ShoppingCart />, color: '#ff9800' },
+                { name: 'Finance', description: 'Banking, Trading, DeFi, Payments', icon: <AttachMoney />, color: '#4caf50' },
+                { name: 'Education', description: 'Learning Platforms, Training, Schools', icon: <School />, color: '#9c27b0' },
+                { name: 'Healthcare', description: 'Medical Apps, Telemedicine, Health', icon: <LocalHospital />, color: '#f44336' },
+                { name: 'Social', description: 'Social Networks, Communities, Chat', icon: <Group />, color: '#00bcd4' },
+                { name: 'Entertainment', description: 'Streaming, Media, Content Creation', icon: <Public />, color: '#ff5722' },
+              ].map((useCase, index) => (
+                <Grid item xs={12} sm={6} md={4} lg={3} key={index}>
+                  <motion.div
+                    variants={itemVariants}
+                    whileHover={{ y: -5, scale: 1.02 }}
+                    transition={{ duration: 0.2 }}
+                  >
+                    <Card sx={{ 
+                      height: '100%', 
+                      p: 2, 
+                      textAlign: 'center',
+                      border: 1,
+                      borderColor: 'divider',
+                      '&:hover': {
+                        borderColor: useCase.color,
+                        boxShadow: `0 4px 20px rgba(0,0,0,0.1)`,
+                      }
+                    }}>
+                      <CardContent>
+                        <Box sx={{ mb: 2, color: useCase.color }}>
+                          {React.cloneElement(useCase.icon, { sx: { fontSize: 40 } })}
+                        </Box>
+                        <Typography variant="h6" gutterBottom sx={{ fontWeight: 'bold' }}>
+                          {useCase.name}
+                        </Typography>
+                        <Typography variant="body2" color="text.secondary">
+                          {useCase.description}
+                        </Typography>
+                      </CardContent>
+                    </Card>
+                  </motion.div>
+                </Grid>
+              ))}
+            </Grid>
+          </motion.div>
+        </Container>
+      </Box>
+
+      {/* STARNET Holons Section */}
+      <Box sx={{ py: 12, bgcolor: 'background.paper' }}>
+        <Container maxWidth="lg">
+          <motion.div variants={itemVariants}>
+            <Typography variant="h4" align="center" gutterBottom sx={{ mb: 8, fontWeight: 'bold' }}>
+              STARNET Holons - The Building Blocks of the Metaverse
+            </Typography>
+            
+            <Typography variant="h6" align="center" color="text.secondary" sx={{ mb: 6 }}>
+              Any STARNETHolon can be linked to any other STARNETHolon as dependencies, creating infinite possibilities for unique combinations
+            </Typography>
+            
+            <Grid container spacing={4}>
+              {[
+                { name: 'OAPPs', description: 'OASIS Applications', icon: <Apps />, color: '#2196f3' },
+                { name: 'Runtimes', description: 'Execution Environments', icon: <Build />, color: '#673ab7' },
+                { name: 'Libraries', description: 'Code Libraries', icon: <LibraryBooks />, color: '#8bc34a' },
+                { name: 'Templates', description: 'Reusable Templates', icon: <Code />, color: '#ff5722' },
+                { name: 'NFTs', description: 'Non-Fungible Tokens', icon: <Image />, color: '#ff9800' },
+                { name: 'GeoNFTs', description: 'Geospatial NFTs', icon: <LocationOn />, color: '#4caf50' },
+                { name: 'GeoHotSpots', description: 'Location-based Hotspots', icon: <LocationOn />, color: '#4caf50' },
+                { name: 'Quests', description: 'Interactive Quests', icon: <Assignment />, color: '#9c27b0' },
+                { name: 'Missions', description: 'Mission Objectives', icon: <FlightTakeoff />, color: '#f44336' },
+                { name: 'Chapters', description: 'Story Chapters', icon: <MenuBook />, color: '#795548' },
+                { name: 'Inventory Items', description: 'Game Items & Rewards', icon: <Inventory />, color: '#607d8b' },
+                { name: 'Celestial Spaces', description: 'MagicVerses, Superverses, Multiverses, Universes, Galaxies, Solar Systems & More!', icon: <Public />, color: '#00bcd4' },
+                { name: 'Celestial Bodies', description: 'Stars, Planets, Moons & More!', icon: <SpaceDashboard />, color: '#3f51b5' },
+                { name: 'Zomes', description: 'Holochain Zomes', icon: <Memory />, color: '#e91e63' },
+                { name: 'Holons', description: 'Basic Data Structures', icon: <DataObject />, color: '#009688' },
+                { name: 'MetaData DNA', description: 'DNA Metadata', icon: <Code />, color: '#ff5722' },
+              ].map((holon, index) => (
+                <Grid item xs={12} sm={6} md={4} lg={3} key={index}>
+                  <motion.div
+                    variants={itemVariants}
+                    whileHover={{ y: -5, scale: 1.02 }}
+                    transition={{ duration: 0.2 }}
+                  >
+                    <Card sx={{ 
+                      height: '100%', 
+                      p: 2, 
+                      textAlign: 'center',
+                      border: 1,
+                      borderColor: 'divider',
+                      '&:hover': {
+                        borderColor: holon.color,
+                        boxShadow: `0 4px 20px rgba(0,0,0,0.1)`,
+                      }
+                    }}>
+                      <CardContent>
+                        <Box sx={{ mb: 2, color: holon.color }}>
+                          {React.cloneElement(holon.icon, { sx: { fontSize: 40 } })}
+                        </Box>
+                        <Typography variant="h6" gutterBottom sx={{ fontWeight: 'bold' }}>
+                          {holon.name}
+                        </Typography>
+                        <Typography variant="body2" color="text.secondary">
+                          {holon.description}
+                        </Typography>
+                      </CardContent>
+                    </Card>
+                  </motion.div>
+                </Grid>
+              ))}
+            </Grid>
+            
+            {/*<Box sx={{ mt: 6, textAlign: 'center' }}>*/}
+            {/*  <Typography variant="h6" gutterBottom>*/}
+            {/*    Example Linkings - Infinite Possibilities:*/}
+            {/*  </Typography>*/}
+            {/*  <Stack direction="row" spacing={2} justifyContent="center" flexWrap="wrap">*/}
+            {/*    <Chip label="Quest + GeoNFT" color="primary" />*/}
+            {/*    <Chip label="OAPP + NFT" color="secondary" />*/}
+            {/*    <Chip label="Mission + Inventory Items" color="success" />*/}
+            {/*    <Chip label="GeoHotSpot + Quest" color="warning" />*/}
+            {/*    <Chip label="Template + Library" color="info" />*/}
+            {/*    <Chip label="Runtime + Zome" color="error" />*/}
+            {/*    <Chip label="Celestial Space + Celestial Body" color="primary" />*/}
+            {/*    <Chip label="Any STARNETHolon + Any STARNETHolon!" color="secondary" />*/}
+            {/*  </Stack>*/}
+            {/*</Box>*/}
+          </motion.div>
+        </Container>
+      </Box>
+
+      {/* Platform Overview */}
+      <Box sx={{ py: 12, bgcolor: 'background.paper' }}>
+        <Container maxWidth="lg">
+          <motion.div variants={itemVariants}>
+            <Typography variant="h4" align="center" gutterBottom sx={{ mb: 8, fontWeight: 'bold' }}>
+              OASIS Architecture Overview
             </Typography>
             
             <Grid container spacing={4}>
@@ -284,20 +500,23 @@ const HomePage: React.FC = () => {
                 <Paper sx={{ p: 4, height: '100%' }}>
                   <Box sx={{ display: 'flex', alignItems: 'center', mb: 3 }}>
                     <Avatar sx={{ bgcolor: 'primary.main', mr: 2 }}>
-                      <Explore />
+                      <Storage />
                     </Avatar>
                     <Typography variant="h5" sx={{ fontWeight: 'bold' }}>
-                      Virtual Worlds
+                      WEB4 OASIS API
                     </Typography>
                   </Box>
                   <Typography variant="body1" color="text.secondary" sx={{ mb: 3 }}>
-                    Explore thousands of virtual worlds, from fantasy realms to futuristic cities. 
-                    Each world offers unique experiences and endless possibilities.
+                    The foundational data aggregation and identity layer with OASIS HyperDrive 
+                    auto-failover/auto-load balance/auto-replication system. Universal data aggregation across all Web2 and Web3 providers.
                   </Typography>
                   <Stack direction="row" spacing={1} flexWrap="wrap">
-                    <Chip label="3D Environments" size="small" />
-                    <Chip label="Social Interaction" size="small" />
-                    <Chip label="Real-time Events" size="small" />
+                    <Chip label="Auto-Failover" size="small" />
+                    <Chip label="Auto-Load Balancing" size="small" />
+                    <Chip label="Auto-Replication" size="small" />
+                    <Chip label="Universal Data" size="small" />
+                    <Chip label="Identity Management" size="small" />
+                    <Chip label="Karma System" size="small" />
                   </Stack>
                 </Paper>
               </Grid>
@@ -306,20 +525,25 @@ const HomePage: React.FC = () => {
                 <Paper sx={{ p: 4, height: '100%' }}>
                   <Box sx={{ display: 'flex', alignItems: 'center', mb: 3 }}>
                     <Avatar sx={{ bgcolor: 'secondary.main', mr: 2 }}>
-                      <Code />
+                      <Apps />
                     </Avatar>
                     <Typography variant="h5" sx={{ fontWeight: 'bold' }}>
-                      Development Tools
+                      WEB5 STAR API
                     </Typography>
                   </Box>
                   <Typography variant="body1" color="text.secondary" sx={{ mb: 3 }}>
-                    Build and deploy applications with our comprehensive development platform. 
-                    From simple scripts to complex AI systems.
+                    The gamification and business layer with STAR ODK (Omniverse Interoperable 
+                    Metaverse Low Code Generator). Build OAPPs, manage quests, missions and create NFTs, Geo-NFTs, Geo-HotSpots & much more!
                   </Typography>
                   <Stack direction="row" spacing={1} flexWrap="wrap">
-                    <Chip label="SDK & APIs" size="small" />
-                    <Chip label="AI Integration" size="small" />
-                    <Chip label="Cloud Deployment" size="small" />
+                    <Chip label="STAR ODK" size="small" />
+                    <Chip label="OAPPs Framework" size="small" />
+                    <Chip label="Libraries, Runtimes, Templates & Pugins" size="small" />
+                    <Chip label="Chaptes, Missions, Quests & Inventory Items" size="small" />
+                    <Chip label="NFTs, Geo-NFTs & Geo-HotSpots" size="small" />
+                    <Chip label="Celestial Spaces & Celestial Bodies" size="small" />
+                    <Chip label="Zomes, Holons & Meta Data DNA" size="small" />
+                    <Chip label="No/Low-Code Development" size="small" />
                   </Stack>
                 </Paper>
               </Grid>
@@ -331,17 +555,18 @@ const HomePage: React.FC = () => {
                       <Store />
                     </Avatar>
                     <Typography variant="h5" sx={{ fontWeight: 'bold' }}>
-                      Marketplace
+                      Provider Ecosystem
                     </Typography>
                   </Box>
                   <Typography variant="body1" color="text.secondary" sx={{ mb: 3 }}>
-                    Buy, sell, and trade digital assets in our secure marketplace. 
-                    From virtual real estate to unique digital collectibles.
+                    Support for 50+ Web2 and Web3 providers including Ethereum, Solana, Holochain, 
+                    IPFS, AWS, Azure, and more. Hot-swappable provider architecture.
                   </Typography>
                   <Stack direction="row" spacing={1} flexWrap="wrap">
-                    <Chip label="NFT Trading" size="small" />
-                    <Chip label="Virtual Real Estate" size="small" />
-                    <Chip label="Digital Art" size="small" />
+                    <Chip label="50+ Providers" size="small" />
+                    <Chip label="Hot-Swappable" size="small" />
+                    <Chip label="Auto-Optimization" size="small" />
+                    <Chip label="Zero Vendor Lock-in" size="small" />
                   </Stack>
                 </Paper>
               </Grid>
@@ -353,17 +578,67 @@ const HomePage: React.FC = () => {
                       <Group />
                     </Avatar>
                     <Typography variant="h5" sx={{ fontWeight: 'bold' }}>
-                      Community
+                      Universal Identity
                     </Typography>
                   </Box>
                   <Typography variant="body1" color="text.secondary" sx={{ mb: 3 }}>
-                    Join millions of users in the largest virtual community. 
-                    Collaborate, learn, and grow together in the OASIS.
+                    Single identity across all platforms with DID support, cross-platform authentication, 
+                    and privacy-preserving identity management with karma integration.
                   </Typography>
                   <Stack direction="row" spacing={1} flexWrap="wrap">
-                    <Chip label="Global Network" size="small" />
-                    <Chip label="Collaboration Tools" size="small" />
-                    <Chip label="Knowledge Sharing" size="small" />
+                    <Chip label="DID Support" size="small" />
+                    <Chip label="Cross-Platform Auth" size="small" />
+                    <Chip label="Privacy-Preserving" size="small" />
+                    <Chip label="Karma Integration" size="small" />
+                    <Chip label="Quad Layer Quantum/Biometric/Aura Encryption System" size="small" />
+                  </Stack>
+                </Paper>
+              </Grid>
+              
+              <Grid item xs={12} md={6}>
+                <Paper sx={{ p: 4, height: '100%' }}>
+                  <Box sx={{ display: 'flex', alignItems: 'center', mb: 3 }}>
+                    <Avatar sx={{ bgcolor: 'info.main', mr: 2 }}>
+                      <Store />
+                    </Avatar>
+                    <Typography variant="h5" sx={{ fontWeight: 'bold' }}>
+                      STARNET Web UI
+                    </Typography>
+                  </Box>
+                  <Typography variant="body1" color="text.secondary" sx={{ mb: 3 }}>
+                    Comprehensive web interface and app store for STAR, STARNET & The OASIS. 
+                    Drag-and-drop OAPP builder, asset management, and publishing platform.
+                  </Typography>
+                  <Stack direction="row" spacing={1} flexWrap="wrap">
+                    <Chip label="App Store" size="small" />
+                    <Chip label="OAPP Builder" size="small" />
+                    <Chip label="Asset Management" size="small" />
+                    <Chip label="STARNET Holon Management" size="small" />
+                    <Chip label="Publishing Platform" size="small" />
+                  </Stack>
+                </Paper>
+              </Grid>
+              
+              <Grid item xs={12} md={6}>
+                <Paper sx={{ p: 4, height: '100%' }}>
+                  <Box sx={{ display: 'flex', alignItems: 'center', mb: 3 }}>
+                    <Avatar sx={{ bgcolor: 'error.main', mr: 2 }}>
+                      <Code />
+                    </Avatar>
+                    <Typography variant="h5" sx={{ fontWeight: 'bold' }}>
+                      STAR CLI & DNA System
+                    </Typography>
+                  </Box>
+                  <Typography variant="body1" color="text.secondary" sx={{ mb: 3 }}>
+                    Powerful command-line interface for developers. Link any STARNET Holon to any other 
+                    STARNE THolon as dependencies. Unlimited combinations and creativity.
+                  </Typography>
+                  <Stack direction="row" spacing={1} flexWrap="wrap">
+                    <Chip label="STAR CLI" size="small" />
+                    <Chip label="DNA/STARNET (Holon) System" size="small" />
+                    <Chip label="Dependencies" size="small" />
+                    <Chip label="Developer Tools" size="small" />
+                    <Chip label="Powerful CLI To The OASIS" size="small" />
                   </Stack>
                 </Paper>
               </Grid>
@@ -378,11 +653,13 @@ const HomePage: React.FC = () => {
           <motion.div variants={itemVariants}>
             <Box sx={{ textAlign: 'center' }}>
               <Typography variant="h4" gutterBottom sx={{ fontWeight: 'bold', mb: 3 }}>
-                Ready to Enter the OASIS?
+                Ready to Build the Future?
               </Typography>
               <Typography variant="h6" sx={{ mb: 6, opacity: 0.9 }}>
-                Join the future of virtual reality and be part of the most advanced 
-                digital ecosystem ever created.
+                Join the OASIS ecosystem and be part of the first universal Web4/Web5 infrastructure 
+                that connects everything to everything. Use STAR CLI and STARNET Web UI to build OAPPs, 
+                link ALL STARNET Holons as dependencies, manage assets, earn karma, and shape the future of 
+                games, businesses, shops, e-commerce, finance, education, healthcare, and everything else!
               </Typography>
               <Stack direction="row" spacing={2} justifyContent="center">
                 <Button
