@@ -1,658 +1,617 @@
+/**
+ * NFTs Page
+ * Complete NFT management interface
+ */
+
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
+import { useDemoMode } from '../contexts/DemoModeContext';
 import {
   Box,
   Typography,
+  Button,
   Card,
   CardContent,
   Grid,
-  Button,
-  TextField,
+  Chip,
+  IconButton,
+  Menu,
+  MenuItem,
   Dialog,
   DialogTitle,
   DialogContent,
   DialogActions,
-  IconButton,
-  Chip,
-  Avatar,
-  Alert,
-  CircularProgress,
-  Badge,
-  Tooltip,
-  LinearProgress,
-  CardMedia,
-  CardActions,
+  TextField,
   FormControl,
   InputLabel,
   Select,
-  MenuItem,
+  Fab,
+  Tooltip,
+  Tabs,
+  Tab,
+  Badge,
+  Stack,
+  Avatar,
+  CardMedia,
+  CardActions,
+  Divider,
 } from '@mui/material';
 import {
-  Add as AddIcon,
-  Refresh as RefreshIcon,
-  Image as ImageIcon,
-  Delete as DeleteIcon,
-  Edit as EditIcon,
-  Visibility as ViewIcon,
-  Star as StarIcon,
-  TrendingUp as TrendingUpIcon,
-  AttachMoney as MoneyIcon,
-  Schedule as ScheduleIcon,
-  FilterList as FilterIcon,
+  Add,
+  MoreVert,
+  PlayArrow,
+  Pause,
+  Download,
+  Upload,
+  Delete,
+  Edit,
+  Visibility,
+  Image,
+  FilterList,
+  Search,
+  Help,
+  Info,
+  Build,
+  MonetizationOn,
+  Share,
+  Favorite,
+  Star,
 } from '@mui/icons-material';
 import { motion } from 'framer-motion';
 import { useQuery, useMutation, useQueryClient } from 'react-query';
-import { starService } from '../services/starService';
-import toast from 'react-hot-toast';
+import { nftService } from '../services';
+import { toast } from 'react-hot-toast';
 
-interface NFT {
-  id: string;
-  name: string;
-  description: string;
-  imageUrl: string;
-  price: number;
-  rarity: 'Common' | 'Rare' | 'Epic' | 'Legendary';
-  category: string;
-  owner: string;
-  createdAt: string;
-  isForSale: boolean;
-  views: number;
-  isInstalled?: boolean; // Added for installed badge and filtering
-  likes: number;
+interface TabPanelProps {
+  children?: React.ReactNode;
+  index: number;
+  value: number;
+}
+
+function TabPanel(props: TabPanelProps) {
+  const { children, value, index, ...other } = props;
+  return (
+    <div
+      role="tabpanel"
+      hidden={value !== index}
+      id={`nft-tabpanel-${index}`}
+      aria-labelledby={`nft-tab-${index}`}
+      {...other}
+    >
+      {value === index && <Box sx={{ p: 3 }}>{children}</Box>}
+    </div>
+  );
 }
 
 const NFTsPage: React.FC = () => {
   const navigate = useNavigate();
+  const { isDemoMode } = useDemoMode();
+  const queryClient = useQueryClient();
+  const [tabValue, setTabValue] = useState(0);
+  const [searchTerm, setSearchTerm] = useState('');
+  const [filterType, setFilterType] = useState('all');
+  const [sortBy, setSortBy] = useState('newest');
   const [createDialogOpen, setCreateDialogOpen] = useState(false);
-  const [filterCategory, setFilterCategory] = useState<string>('all');
-  const [viewScope, setViewScope] = useState<string>('all');
-  const [newNFT, setNewNFT] = useState({
-    name: '',
-    description: '',
-    price: 0,
-    rarity: 'Common' as const,
-    category: 'Art',
+  const [selectedNFT, setSelectedNFT] = useState<any>(null);
+  const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
+
+  // Fetch NFTs
+  const { data: nfts, isLoading, error } = useQuery('nfts', () => nftService.getAll());
+
+  // Create NFT mutation
+  const createNFTMutation = useMutation(
+    async (nftData: any) => {
+      const response = await nftService.create(nftData);
+      return response.result;
+    },
+    {
+      onSuccess: () => {
+        queryClient.invalidateQueries('nfts');
+        toast.success('NFT created successfully!');
+        setCreateDialogOpen(false);
+      },
+      onError: (error: any) => {
+        toast.error('Failed to create NFT: ' + error.message);
+      },
+    }
+  );
+
+  // Publish NFT mutation
+  const publishNFTMutation = useMutation(
+    async (nftId: string) => {
+      const response = await nftService.publish(nftId, {});
+      return response.result;
+    },
+    {
+      onSuccess: () => {
+        queryClient.invalidateQueries('nfts');
+        toast.success('NFT published successfully!');
+      },
+      onError: (error: any) => {
+        toast.error('Failed to publish NFT: ' + error.message);
+      },
+    }
+  );
+
+  // Download NFT mutation
+  const downloadNFTMutation = useMutation(
+    async (nftId: string) => {
+      const response = await nftService.download(nftId, './downloads', true);
+      return response.result;
+    },
+    {
+      onSuccess: () => {
+        toast.success('NFT downloaded successfully!');
+      },
+      onError: (error: any) => {
+        toast.error('Failed to download NFT: ' + error.message);
+      },
+    }
+  );
+
+  const handleCreateNFT = (nftData: any) => {
+    createNFTMutation.mutate(nftData);
+  };
+
+  const handlePublishNFT = (nftId: string) => {
+    publishNFTMutation.mutate(nftId);
+  };
+
+  const handleDownloadNFT = (nftId: string) => {
+    downloadNFTMutation.mutate(nftId);
+  };
+
+  const handleMenuClick = (event: React.MouseEvent<HTMLElement>, nft: any) => {
+    setAnchorEl(event.currentTarget);
+    setSelectedNFT(nft);
+  };
+
+  const handleMenuClose = () => {
+    setAnchorEl(null);
+    setSelectedNFT(null);
+  };
+
+  const filteredNFTs = nfts?.result?.filter((nft: any) => {
+    const matchesSearch = nft.name?.toLowerCase().includes(searchTerm.toLowerCase()) ||
+                         nft.description?.toLowerCase().includes(searchTerm.toLowerCase());
+    const matchesFilter = filterType === 'all' || nft.type === filterType;
+    return matchesSearch && matchesFilter;
+  }) || [];
+
+  const sortedNFTs = [...filteredNFTs].sort((a: any, b: any) => {
+    switch (sortBy) {
+      case 'newest':
+        return new Date(b.createdOn || 0).getTime() - new Date(a.createdOn || 0).getTime();
+      case 'oldest':
+        return new Date(a.createdOn || 0).getTime() - new Date(b.createdOn || 0).getTime();
+      case 'name':
+        return (a.name || '').localeCompare(b.name || '');
+      case 'value':
+        return (b.value || 0) - (a.value || 0);
+      default:
+        return 0;
+    }
   });
 
-  const queryClient = useQueryClient();
-
-  // Fetch NFTs with impressive demo data
-  const { data: nftsData, isLoading, error, refetch } = useQuery(
-    'nfts',
-    async () => {
-      try {
-        // Force demo data for now
-        throw 'Forcing demo data for NFTs';
-        const response = await starService.getAllNFTs();
-        return response;
-      } catch (error) {
-        // Fallback to impressive demo data
-        console.log('Using demo NFT data for investor presentation');
-        return {
-          result: [
-            {
-              id: '1',
-              name: 'Cosmic Dragon',
-              description: 'A legendary dragon from the depths of the OASIS universe',
-              imageUrl: 'https://images.unsplash.com/photo-1578662996442-48f60103fc96?w=400&h=400&fit=crop',
-              price: 2.5,
-              rarity: 'Legendary',
-              category: 'Creatures',
-              owner: 'OASIS_Explorer',
-              createdAt: '2024-01-15',
-              isForSale: true,
-              views: 1250,
-              likes: 89,
-            },
-            {
-              id: '2',
-              name: 'Neon Cityscape',
-              description: 'A futuristic cityscape from the cyberpunk realm',
-              imageUrl: 'https://images.unsplash.com/photo-1446776877081-d282a0f896e2?w=400&h=400&fit=crop&auto=format&q=80',
-              price: 1.8,
-              rarity: 'Epic',
-              category: 'Art',
-              owner: 'DigitalArtist_99',
-              createdAt: '2024-01-14',
-              isForSale: true,
-              views: 890,
-              likes: 67,
-            },
-            {
-              id: '3',
-              name: 'Quantum Crystal',
-              description: 'A rare crystal that bends reality itself',
-              imageUrl: 'https://images.unsplash.com/photo-1518709268805-4e9042af2176?w=400&h=400&fit=crop',
-              price: 3.2,
-              rarity: 'Legendary',
-              category: 'Minerals',
-              owner: 'QuantumMiner',
-              createdAt: '2024-01-13',
-              isForSale: false,
-              views: 2100,
-              likes: 156,
-            },
-            {
-              id: '4',
-              name: 'Virtual Pet - Cyber Cat',
-              description: 'An adorable cybernetic cat companion',
-              imageUrl: 'https://images.unsplash.com/photo-1574158622682-e40e69881006?w=400&h=400&fit=crop&auto=format&q=80',
-              price: 0.8,
-              rarity: 'Rare',
-              category: 'Pets',
-              owner: 'PetLover_42',
-              createdAt: '2024-01-12',
-              isForSale: true,
-              views: 650,
-              likes: 45,
-            },
-            {
-              id: '5',
-              name: 'Space Station Alpha',
-              description: 'A massive space station orbiting a distant star',
-              imageUrl: 'https://images.unsplash.com/photo-1502134249126-9f3755a50d78?w=400&h=400&fit=crop&auto=format&q=80',
-              price: 4.5,
-              rarity: 'Legendary',
-              category: 'Structures',
-              owner: 'SpaceArchitect',
-              createdAt: '2024-01-11',
-              isForSale: true,
-              views: 1800,
-              likes: 134,
-            },
-            {
-              id: '6',
-              name: 'Holographic Flower',
-              description: 'A beautiful flower that exists only in digital space',
-              imageUrl: 'https://images.unsplash.com/photo-1490750967868-88aa4486c946?w=400&h=400&fit=crop',
-              price: 0.5,
-              rarity: 'Common',
-              category: 'Nature',
-              owner: 'NatureLover',
-              createdAt: '2024-01-10',
-              isForSale: true,
-              views: 320,
-              likes: 23,
-            },
-          ]
-        };
-      }
-    },
-    {
-      refetchInterval: 30000, // Refetch every 30 seconds
-      refetchOnWindowFocus: true,
-    }
-  );
-
-  const createNFTMutation = useMutation(
-    async (nftData: Partial<NFT>) => {
-      try {
-        return await starService.createNFT({
-          name: nftData.name || '',
-          description: nftData.description || '',
-          imageUrl: nftData.imageUrl || '',
-          price: nftData.price || 0,
-          rarity: nftData.rarity || 'Common',
-          category: nftData.category || 'Art'
-        });
-      } catch (error) {
-        // For demo purposes, simulate success
-        toast.success('NFT created successfully! (Demo Mode)');
-        return { success: true, id: Date.now().toString() };
-      }
-    },
-    {
-      onSuccess: () => {
-        queryClient.invalidateQueries('nfts');
-        setCreateDialogOpen(false);
-        setNewNFT({ name: '', description: '', price: 0, rarity: 'Common', category: 'Art' });
-      },
-    }
-  );
-
-  const deleteNFTMutation = useMutation(
-    async (nftId: string) => {
-      try {
-        return await starService.deleteNFT(nftId);
-      } catch (error) {
-        // For demo purposes, simulate success
-        toast.success('NFT deleted successfully! (Demo Mode)');
-        return { success: true };
-      }
-    },
-    {
-      onSuccess: () => {
-        queryClient.invalidateQueries('nfts');
-      },
-    }
-  );
-
-  const handleCreateNFT = () => {
-    if (!newNFT.name || !newNFT.description) {
-      toast.error('Please fill in all required fields');
-      return;
-    }
-    createNFTMutation.mutate(newNFT);
-  };
-
-  const handleDeleteNFT = (nftId: string) => {
-    if (window.confirm('Are you sure you want to delete this NFT?')) {
-      deleteNFTMutation.mutate(nftId);
-    }
-  };
-
-  const getRarityColor = (rarity: string) => {
-    switch (rarity) {
-      case 'Common': return '#9e9e9e';
-      case 'Rare': return '#2196f3';
-      case 'Epic': return '#9c27b0';
-      case 'Legendary': return '#ff9800';
-      default: return '#9e9e9e';
-    }
-  };
-
-  const getCategoryColor = (category: string) => {
-    switch (category.toLowerCase()) {
-      case 'art': return '#e91e63';
-      case 'creatures': return '#4caf50';
-      case 'minerals': return '#9c27b0';
-      case 'pets': return '#ff9800';
-      case 'structures': return '#607d8b';
-      case 'nature': return '#8bc34a';
-      default: return '#1976d2';
-    }
-  };
-
-  const categories = ['all', 'Art', 'Creatures', 'Minerals', 'Pets', 'Structures', 'Nature'];
-  
-  // Filter by view scope first, then by category
-  const getFilteredNFTs = () => {
-    let filtered = nftsData?.result || [];
-    
-    // Apply view scope filter
-    if (viewScope === 'installed') {
-      filtered = filtered.filter((nft: any) => nft.isInstalled);
-    } else if (viewScope === 'mine') {
-      // For demo, show NFTs created by current user
-      filtered = filtered.filter((nft: any) => nft.owner === 'OASIS_Explorer');
-    }
-    
-    // Apply category filter
-    if (filterCategory !== 'all') {
-      filtered = filtered.filter((nft: any) => nft.category === filterCategory);
-    }
-    
-    return filtered;
-  };
-  
-  const filteredNFTs = getFilteredNFTs();
-
-  const containerVariants = {
-    hidden: { opacity: 0 },
-    visible: {
-      opacity: 1,
-      transition: {
-        staggerChildren: 0.1,
-      },
-    },
-  };
-
-  const itemVariants = {
-    hidden: { opacity: 0, y: 20 },
-    visible: { opacity: 1, y: 0 },
+  const nftStats = {
+    total: nfts?.result?.length || 0,
+    published: nfts?.result?.filter((nft: any) => nft.isPublished).length || 0,
+    totalValue: nfts?.result?.reduce((sum: number, nft: any) => sum + (nft.value || 0), 0) || 0,
+    averageRating: nfts?.result?.reduce((sum: number, nft: any) => sum + (nft.rating || 0), 0) / (nfts?.result?.length || 1) || 0,
   };
 
   return (
-    <>
-        <Box sx={{ mb: 4, mt: 4 }}>
-        <Typography variant="h4" gutterBottom className="page-heading">
-          🎨 NFT Marketplace
-        </Typography>
-        <Typography variant="subtitle1" color="text.secondary">
-          Discover, create, and trade digital assets in the OASIS
-        </Typography>
-      </Box>
-
-      {error && (
-        <Alert severity="error" sx={{ mb: 3 }}>
-          Failed to load NFTs. Using demo data for presentation.
-        </Alert>
-      )}
-
-      {/* Action Bar */}
-      <Box sx={{ 
-        display: 'flex', 
-        justifyContent: 'space-between', 
-        alignItems: 'center', 
-        mb: 3,
-        flexWrap: 'wrap',
-        gap: 2
-      }}>
-        <Box sx={{ display: 'flex', gap: 2, alignItems: 'center' }}>
-          <Button
-            variant="contained"
-            startIcon={<AddIcon />}
-            onClick={() => setCreateDialogOpen(true)}
-            sx={{
-              bgcolor: '#1976d2',
-              '&:hover': {
-                bgcolor: '#1565c0',
-              }
-            }}
-          >
-            Create NFT
-          </Button>
-          
-          <Button
-            variant="outlined"
-            startIcon={<RefreshIcon />}
-            onClick={() => refetch()}
-            disabled={isLoading}
-          >
-            Refresh
-          </Button>
-        </Box>
-
-        <Box sx={{ display: 'flex', gap: 2, alignItems: 'center' }}>
-          <FormControl size="small" sx={{ minWidth: 140 }}>
-            <InputLabel>View</InputLabel>
-            <Select
-              value={viewScope}
-              label="View"
-              onChange={(e) => setViewScope(e.target.value)}
-            >
-              <MenuItem value="all">All</MenuItem>
-              <MenuItem value="installed">Installed</MenuItem>
-              <MenuItem value="mine">My NFTs</MenuItem>
-            </Select>
-          </FormControl>
-          
-          <Box sx={{ display: 'flex', gap: 1, alignItems: 'center' }}>
-            <FilterIcon sx={{ color: 'text.secondary' }} />
-            {categories.map((category) => (
-              <Chip
-                key={category}
-                label={category}
-                onClick={() => setFilterCategory(category)}
-                variant={filterCategory === category ? 'filled' : 'outlined'}
-                sx={{
-                  bgcolor: filterCategory === category ? 'primary.main' : 'transparent',
-                  color: filterCategory === category ? 'white' : 'text.primary',
-                  '&:hover': {
-                    bgcolor: filterCategory === category ? 'primary.dark' : 'action.hover',
-                  }
-                }}
-              />
-            ))}
+    <Box sx={{ flexGrow: 1, p: 3 }}>
+      {/* Header */}
+      <Box sx={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', mb: 3 }}>
+        <Box>
+          <Typography variant="h4" component="h1" gutterBottom>
+            NFTs
+          </Typography>
+          <Typography variant="body1" color="text.secondary">
+            Manage your Non-Fungible Tokens
+          </Typography>
+          <Box sx={{ mt: 1, p: 2, bgcolor: '#0d47a1', color: 'white', borderRadius: 1, display: 'flex', alignItems: 'center', gap: 1 }}>
+            <Info sx={{ color: 'white' }} />
+            <Typography variant="body2" sx={{ color: 'white' }}>
+              Create, manage and publish NFTs. Data updates in real-time.
+            </Typography>
           </Box>
         </Box>
+        <Button
+          variant="contained"
+          startIcon={<Add />}
+          onClick={() => setCreateDialogOpen(true)}
+          sx={{ borderRadius: 2 }}
+        >
+          Create NFT
+        </Button>
       </Box>
 
-      {isLoading && (
-        <Box sx={{ mb: 3 }}>
-          <LinearProgress sx={{ height: 6, borderRadius: 3 }} />
-        </Box>
-      )}
-
-      {/* NFT Grid */}
-      <Grid container spacing={3}>
-        {filteredNFTs.map((nft: any, index: number) => (
-          <Grid item xs={12} sm={6} md={4} lg={3} key={nft.id}>
-            <motion.div
-              variants={itemVariants}
-              whileHover={{ 
-                scale: 1.05,
-                rotateY: 5,
-                transition: { duration: 0.2 }
-              }}
-            >
-              <Card 
-                sx={{ 
-                  height: '100%',
-                  display: 'flex',
-                  flexDirection: 'column',
-                  background: `linear-gradient(135deg, ${getRarityColor(nft.rarity)}15, ${getRarityColor(nft.rarity)}05)`,
-                  border: `2px solid ${getRarityColor(nft.rarity)}30`,
-                  boxShadow: `0 8px 32px ${getRarityColor(nft.rarity)}20`,
-                  transition: 'all 0.3s ease',
-                  cursor: 'pointer',
-                  '&:hover': {
-                    boxShadow: `0 12px 40px ${getRarityColor(nft.rarity)}30`,
-                  }
-                }}
-                onClick={() => navigate(`/nfts/${nft.id}`)}
-              >
-                <Box sx={{ position: 'relative' }}>
-                  <Box
-                    sx={{
-                      height: 200,
-                      backgroundImage: `url(${nft.imageUrl})`,
-                      backgroundSize: 'cover',
-                      backgroundPosition: 'center',
-                      backgroundRepeat: 'no-repeat',
-                      filter: 'brightness(1.1) contrast(1.1)',
-                    }}
-                  />
-                  <Chip
-                    label={nft.category}
-                    size="small"
-                    sx={{
-                      position: 'absolute',
-                      top: 8,
-                      left: 8,
-                      bgcolor: getCategoryColor(nft.category),
-                      color: 'white',
-                      fontWeight: 'bold',
-                      fontSize: '0.7rem'
-                    }}
-                  />
-                  <Chip
-                    label={nft.rarity}
-                    size="small"
-                    sx={{
-                      position: 'absolute',
-                      top: 8,
-                      right: 8,
-                      bgcolor: getRarityColor(nft.rarity),
-                      color: 'white',
-                      fontWeight: 'bold',
-                      fontSize: '0.7rem'
-                    }}
-                  />
-                  {nft.isForSale && (
-                    <Chip
-                      label="FOR SALE"
-                      size="small"
-                      sx={{
-                        position: 'absolute',
-                        top: 40,
-                        left: 8,
-                        bgcolor: '#4caf50',
-                        color: 'white',
-                        fontWeight: 'bold',
-                        fontSize: '0.7rem'
-                      }}
-                    />
-                  )}
-                  {nft.isInstalled && (
-                    <Chip
-                      label="Installed"
-                      size="small"
-                      color="success"
-                      sx={{
-                        position: 'absolute',
-                        bottom: 8,
-                        right: 8,
-                        fontWeight: 'bold',
-                      }}
-                    />
-                  )}
+      {/* Stats Cards */}
+      <Grid container spacing={3} sx={{ mb: 3 }}>
+        <Grid item xs={12} sm={6} md={3}>
+          <Card>
+            <CardContent>
+              <Box sx={{ display: 'flex', alignItems: 'center' }}>
+                <Image color="primary" sx={{ mr: 2 }} />
+                <Box>
+                  <Typography variant="h6">{nftStats.total}</Typography>
+                  <Typography variant="body2" color="text.secondary">
+                    Total NFTs
+                  </Typography>
                 </Box>
-
-                <CardContent sx={{ flexGrow: 1 }}>
-                  <Typography variant="h6" gutterBottom sx={{ fontWeight: 'bold' }}>
-                    {nft.name}
+              </Box>
+            </CardContent>
+          </Card>
+        </Grid>
+        <Grid item xs={12} sm={6} md={3}>
+          <Card>
+            <CardContent>
+              <Box sx={{ display: 'flex', alignItems: 'center' }}>
+                <Upload color="success" sx={{ mr: 2 }} />
+                <Box>
+                  <Typography variant="h6">{nftStats.published}</Typography>
+                  <Typography variant="body2" color="text.secondary">
+                    Published
                   </Typography>
-                  
-                  <Typography variant="body2" color="text.secondary" sx={{ mb: 2 }}>
-                    {nft.description}
+                </Box>
+              </Box>
+            </CardContent>
+          </Card>
+        </Grid>
+        <Grid item xs={12} sm={6} md={3}>
+          <Card>
+            <CardContent>
+              <Box sx={{ display: 'flex', alignItems: 'center' }}>
+                <MonetizationOn color="warning" sx={{ mr: 2 }} />
+                <Box>
+                  <Typography variant="h6">{nftStats.totalValue.toFixed(2)} ETH</Typography>
+                  <Typography variant="body2" color="text.secondary">
+                    Total Value
                   </Typography>
-
-                  <Box sx={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', mb: 2 }}>
-                    <Box sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
-                      <MoneyIcon sx={{ color: '#4caf50', fontSize: 20 }} />
-                      <Typography variant="h6" sx={{ color: '#4caf50', fontWeight: 'bold' }}>
-                        {nft.price} ETH
-                      </Typography>
-                    </Box>
-                    <Chip
-                      label={nft.category}
-                      size="small"
-                      variant="outlined"
-                    />
-                  </Box>
-
-                  <Box sx={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}>
-                    <Box sx={{ display: 'flex', gap: 2 }}>
-                      <Tooltip title="Views">
-                        <Box sx={{ display: 'flex', alignItems: 'center', gap: 0.5 }}>
-                          <ViewIcon sx={{ fontSize: 16, color: 'text.secondary' }} />
-                          <Typography variant="caption">{nft.views}</Typography>
-                        </Box>
-                      </Tooltip>
-                      <Tooltip title="Likes">
-                        <Box sx={{ display: 'flex', alignItems: 'center', gap: 0.5 }}>
-                          <StarIcon sx={{ fontSize: 16, color: '#ff9800' }} />
-                          <Typography variant="caption">{nft.likes}</Typography>
-                        </Box>
-                      </Tooltip>
-                    </Box>
-                    <Typography variant="caption" color="text.secondary">
-                      by {nft.owner}
-                    </Typography>
-                  </Box>
-                </CardContent>
-
-                <CardActions sx={{ justifyContent: 'space-between', p: 2 }}>
-                  <Button
-                    size="small"
-                    startIcon={<ViewIcon />}
-                    sx={{ color: 'primary.main' }}
-                  >
-                    View
-                  </Button>
-                  <Box>
-                    <IconButton
-                      size="small"
-                      onClick={() => handleDeleteNFT(nft.id)}
-                      disabled={deleteNFTMutation.isLoading}
-                      sx={{ color: 'error.main' }}
-                    >
-                      <DeleteIcon />
-                    </IconButton>
-                  </Box>
-                </CardActions>
-              </Card>
-            </motion.div>
-          </Grid>
-        ))}
+                </Box>
+              </Box>
+            </CardContent>
+          </Card>
+        </Grid>
+        <Grid item xs={12} sm={6} md={3}>
+          <Card>
+            <CardContent>
+              <Box sx={{ display: 'flex', alignItems: 'center' }}>
+                <Star color="info" sx={{ mr: 2 }} />
+                <Box>
+                  <Typography variant="h6">{nftStats.averageRating.toFixed(1)}</Typography>
+                  <Typography variant="body2" color="text.secondary">
+                    Avg Rating
+                  </Typography>
+                </Box>
+              </Box>
+            </CardContent>
+          </Card>
+        </Grid>
       </Grid>
 
-      {filteredNFTs.length === 0 && !isLoading && (
-        <Card sx={{ textAlign: 'center', py: 8 }}>
-          <CardContent>
-            <ImageIcon sx={{ fontSize: 80, color: 'text.secondary', mb: 2 }} />
-            <Typography variant="h6" color="text.secondary">
-              No NFTs found
-            </Typography>
-            <Typography variant="body2" color="text.secondary">
-              Try adjusting your filters or create a new NFT
-            </Typography>
-          </CardContent>
-        </Card>
-      )}
+      {/* Filters and Search */}
+      <Card sx={{ mb: 3 }}>
+        <CardContent>
+          <Grid container spacing={2} alignItems="center">
+            <Grid item xs={12} md={4}>
+              <TextField
+                fullWidth
+                placeholder="Search NFTs..."
+                value={searchTerm}
+                onChange={(e) => setSearchTerm(e.target.value)}
+                InputProps={{
+                  startAdornment: <Search sx={{ mr: 1, color: 'text.secondary' }} />,
+                }}
+              />
+            </Grid>
+            <Grid item xs={12} md={3}>
+              <FormControl fullWidth>
+                <InputLabel>Filter by Type</InputLabel>
+                <Select
+                  value={filterType}
+                  onChange={(e) => setFilterType(e.target.value)}
+                >
+                  <MenuItem value="all">All Types</MenuItem>
+                  <MenuItem value="Art">Art</MenuItem>
+                  <MenuItem value="Music">Music</MenuItem>
+                  <MenuItem value="Video">Video</MenuItem>
+                  <MenuItem value="Game">Game</MenuItem>
+                  <MenuItem value="Collectible">Collectible</MenuItem>
+                </Select>
+              </FormControl>
+            </Grid>
+            <Grid item xs={12} md={3}>
+              <FormControl fullWidth>
+                <InputLabel>Sort by</InputLabel>
+                <Select
+                  value={sortBy}
+                  onChange={(e) => setSortBy(e.target.value)}
+                >
+                  <MenuItem value="newest">Newest</MenuItem>
+                  <MenuItem value="oldest">Oldest</MenuItem>
+                  <MenuItem value="name">Name</MenuItem>
+                  <MenuItem value="value">Value</MenuItem>
+                </Select>
+              </FormControl>
+            </Grid>
+            <Grid item xs={12} md={2}>
+              <Button
+                fullWidth
+                variant="outlined"
+                startIcon={<FilterList />}
+              >
+                More Filters
+              </Button>
+            </Grid>
+          </Grid>
+        </CardContent>
+      </Card>
+
+      {/* NFTs Grid */}
+      <Grid container spacing={3}>
+        {isLoading ? (
+          Array.from({ length: 6 }).map((_, index) => (
+            <Grid item xs={12} sm={6} md={4} key={index}>
+              <Card>
+                <CardContent>
+                  <Box sx={{ display: 'flex', alignItems: 'center', mb: 2 }}>
+                    <Avatar sx={{ mr: 2 }} />
+                    <Box sx={{ flexGrow: 1 }}>
+                      <Typography variant="h6" sx={{ mb: 1 }}>
+                        Loading...
+                      </Typography>
+                      <Typography variant="body2" color="text.secondary">
+                        Loading NFT details...
+                      </Typography>
+                    </Box>
+                  </Box>
+                </CardContent>
+              </Card>
+            </Grid>
+          ))
+        ) : error ? (
+          <Grid item xs={12}>
+            <Card>
+              <CardContent sx={{ textAlign: 'center', py: 4 }}>
+                <Typography variant="h6" color="error" gutterBottom>
+                  Failed to load NFTs
+                </Typography>
+                <Typography variant="body2" color="text.secondary">
+                  {error instanceof Error ? error.message : 'An error occurred'}
+                </Typography>
+              </CardContent>
+            </Card>
+          </Grid>
+        ) : sortedNFTs.length === 0 ? (
+          <Grid item xs={12}>
+            <Card>
+              <CardContent sx={{ textAlign: 'center', py: 4 }}>
+                <Image sx={{ fontSize: 64, color: 'text.secondary', mb: 2 }} />
+                <Typography variant="h6" gutterBottom>
+                  No NFTs found
+                </Typography>
+                <Typography variant="body2" color="text.secondary" sx={{ mb: 3 }}>
+                  {searchTerm ? 'Try adjusting your search criteria' : 'Create your first NFT to get started'}
+                </Typography>
+                <Button
+                  variant="contained"
+                  startIcon={<Add />}
+                  onClick={() => setCreateDialogOpen(true)}
+                >
+                  Create NFT
+                </Button>
+              </CardContent>
+            </Card>
+          </Grid>
+        ) : (
+          sortedNFTs.map((nft: any) => (
+            <Grid item xs={12} sm={6} md={4} key={nft.id}>
+              <motion.div
+                initial={{ opacity: 0, y: 20 }}
+                animate={{ opacity: 1, y: 0 }}
+                transition={{ duration: 0.3 }}
+              >
+                <Card sx={{ height: '100%', display: 'flex', flexDirection: 'column' }}>
+                  <CardMedia
+                    component="img"
+                    height="200"
+                    image={nft.imageUrl || '/api/placeholder/400/200'}
+                    alt={nft.name}
+                    sx={{ objectFit: 'cover' }}
+                  />
+                  <CardContent sx={{ flexGrow: 1 }}>
+                    <Box sx={{ display: 'flex', justifyContent: 'space-between', alignItems: 'flex-start', mb: 2 }}>
+                      <Typography variant="h6" component="h3" noWrap>
+                        {nft.name}
+                      </Typography>
+                      <IconButton
+                        size="small"
+                        onClick={(e) => handleMenuClick(e, nft)}
+                      >
+                        <MoreVert />
+                      </IconButton>
+                    </Box>
+                    <Typography variant="body2" color="text.secondary" sx={{ mb: 2 }}>
+                      {nft.description}
+                    </Typography>
+                    <Box sx={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', mb: 2 }}>
+                      <Chip
+                        label={nft.type || 'NFT'}
+                        size="small"
+                        color="primary"
+                        variant="outlined"
+                      />
+                      <Box sx={{ display: 'flex', alignItems: 'center' }}>
+                        <Star sx={{ fontSize: 16, mr: 0.5, color: 'warning.main' }} />
+                        <Typography variant="body2">
+                          {nft.rating || 0}
+                        </Typography>
+                      </Box>
+                    </Box>
+                    <Box sx={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}>
+                      <Typography variant="h6" color="primary">
+                        {nft.value || '0'} ETH
+                      </Typography>
+                      <Box sx={{ display: 'flex', alignItems: 'center' }}>
+                        <Typography variant="body2" color="text.secondary" sx={{ mr: 1 }}>
+                          {nft.downloads || 0} downloads
+                        </Typography>
+                      </Box>
+                    </Box>
+                  </CardContent>
+                  <Divider />
+                  <CardActions sx={{ justifyContent: 'space-between', p: 2 }}>
+                    <Button
+                      size="small"
+                      startIcon={<Visibility />}
+                      onClick={() => navigate(`/nfts/${nft.id}`)}
+                    >
+                      View
+                    </Button>
+                    <Box>
+                      <IconButton
+                        size="small"
+                        onClick={() => handleDownloadNFT(nft.id)}
+                        disabled={downloadNFTMutation.isLoading}
+                      >
+                        <Download />
+                      </IconButton>
+                      <IconButton
+                        size="small"
+                        onClick={() => handlePublishNFT(nft.id)}
+                        disabled={publishNFTMutation.isLoading}
+                      >
+                        <Upload />
+                      </IconButton>
+                    </Box>
+                  </CardActions>
+                </Card>
+              </motion.div>
+            </Grid>
+          ))
+        )}
+      </Grid>
 
       {/* Create NFT Dialog */}
-      <Dialog open={createDialogOpen} onClose={() => setCreateDialogOpen(false)} maxWidth="sm" fullWidth>
+      <Dialog
+        open={createDialogOpen}
+        onClose={() => setCreateDialogOpen(false)}
+        maxWidth="md"
+        fullWidth
+      >
         <DialogTitle>Create New NFT</DialogTitle>
         <DialogContent>
-          <TextField
-            autoFocus
-            margin="dense"
-            label="NFT Name"
-            fullWidth
-            variant="outlined"
-            value={newNFT.name}
-            onChange={(e) => setNewNFT({ ...newNFT, name: e.target.value })}
-            sx={{ mb: 2 }}
-          />
-          <TextField
-            margin="dense"
-            label="Description"
-            fullWidth
-            multiline
-            rows={3}
-            variant="outlined"
-            value={newNFT.description}
-            onChange={(e) => setNewNFT({ ...newNFT, description: e.target.value })}
-            sx={{ mb: 2 }}
-          />
-          <TextField
-            margin="dense"
-            label="Price (ETH)"
-            type="number"
-            fullWidth
-            variant="outlined"
-            value={newNFT.price}
-            onChange={(e) => setNewNFT({ ...newNFT, price: parseFloat(e.target.value) || 0 })}
-            sx={{ mb: 2 }}
-          />
-          <TextField
-            margin="dense"
-            label="Category"
-            fullWidth
-            variant="outlined"
-            value={newNFT.category}
-            onChange={(e) => setNewNFT({ ...newNFT, category: e.target.value })}
-            select
-            SelectProps={{ native: true }}
-            sx={{ mb: 2 }}
-          >
-            <option value="Art">Art</option>
-            <option value="Creatures">Creatures</option>
-            <option value="Minerals">Minerals</option>
-            <option value="Pets">Pets</option>
-            <option value="Structures">Structures</option>
-            <option value="Nature">Nature</option>
-            <option value="Landscapes">Landscapes</option>
-            <option value="Characters">Characters</option>
-            <option value="Vehicles">Vehicles</option>
-            <option value="Weapons">Weapons</option>
-            <option value="Accessories">Accessories</option>
-            <option value="Music">Music</option>
-            <option value="Sports">Sports</option>
-            <option value="Gaming">Gaming</option>
-            <option value="Technology">Technology</option>
-            <option value="Fashion">Fashion</option>
-            <option value="Collectibles">Collectibles</option>
-            <option value="Memes">Memes</option>
-            <option value="Abstract">Abstract</option>
-            <option value="Photography">Photography</option>
-          </TextField>
+          <Grid container spacing={2} sx={{ mt: 1 }}>
+            <Grid item xs={12}>
+              <TextField
+                fullWidth
+                label="NFT Name"
+                placeholder="Enter NFT name"
+                required
+              />
+            </Grid>
+            <Grid item xs={12}>
+              <TextField
+                fullWidth
+                label="Description"
+                placeholder="Enter NFT description"
+                multiline
+                rows={3}
+              />
+            </Grid>
+            <Grid item xs={12} sm={6}>
+              <FormControl fullWidth>
+                <InputLabel>Type</InputLabel>
+                <Select defaultValue="Art">
+                  <MenuItem value="Art">Art</MenuItem>
+                  <MenuItem value="Music">Music</MenuItem>
+                  <MenuItem value="Video">Video</MenuItem>
+                  <MenuItem value="Game">Game</MenuItem>
+                  <MenuItem value="Collectible">Collectible</MenuItem>
+                </Select>
+              </FormControl>
+            </Grid>
+            <Grid item xs={12} sm={6}>
+              <TextField
+                fullWidth
+                label="Value (ETH)"
+                type="number"
+                placeholder="0.1"
+              />
+            </Grid>
+            <Grid item xs={12}>
+              <TextField
+                fullWidth
+                label="Image URL"
+                placeholder="https://example.com/image.png"
+              />
+            </Grid>
+          </Grid>
         </DialogContent>
         <DialogActions>
-          <Button onClick={() => setCreateDialogOpen(false)}>Cancel</Button>
+          <Button onClick={() => setCreateDialogOpen(false)}>
+            Cancel
+          </Button>
           <Button
-            onClick={handleCreateNFT}
             variant="contained"
+            onClick={() => handleCreateNFT({})}
             disabled={createNFTMutation.isLoading}
-            startIcon={createNFTMutation.isLoading ? <CircularProgress size={20} /> : <AddIcon />}
           >
             {createNFTMutation.isLoading ? 'Creating...' : 'Create NFT'}
           </Button>
         </DialogActions>
       </Dialog>
-    </>
+
+      {/* Context Menu */}
+      <Menu
+        anchorEl={anchorEl}
+        open={Boolean(anchorEl)}
+        onClose={handleMenuClose}
+      >
+        <MenuItem onClick={() => {
+          if (selectedNFT) navigate(`/nfts/${selectedNFT.id}`);
+          handleMenuClose();
+        }}>
+          <Visibility sx={{ mr: 1 }} />
+          View Details
+        </MenuItem>
+        <MenuItem onClick={() => {
+          if (selectedNFT) handleDownloadNFT(selectedNFT.id);
+          handleMenuClose();
+        }}>
+          <Download sx={{ mr: 1 }} />
+          Download
+        </MenuItem>
+        <MenuItem onClick={() => {
+          if (selectedNFT) handlePublishNFT(selectedNFT.id);
+          handleMenuClose();
+        }}>
+          <Upload sx={{ mr: 1 }} />
+          Publish
+        </MenuItem>
+        <MenuItem onClick={handleMenuClose}>
+          <Edit sx={{ mr: 1 }} />
+          Edit
+        </MenuItem>
+        <MenuItem onClick={handleMenuClose} sx={{ color: 'error.main' }}>
+          <Delete sx={{ mr: 1 }} />
+          Delete
+        </MenuItem>
+      </Menu>
+
+      {/* Floating Action Button */}
+      <Fab
+        color="primary"
+        aria-label="create nft"
+        sx={{ position: 'fixed', bottom: 16, right: 16 }}
+        onClick={() => setCreateDialogOpen(true)}
+      >
+        <Add />
+      </Fab>
+    </Box>
   );
 };
 
