@@ -40,7 +40,7 @@ import {
 import { motion } from 'framer-motion';
 import { useQuery, useMutation, useQueryClient } from 'react-query';
 import toast from 'react-hot-toast';
-import { starService } from '../services/starService';
+import { pluginService } from '../services';
 
 interface Plugin {
   id: string;
@@ -93,7 +93,7 @@ const PluginsPage: React.FC = () => {
     async () => {
       try {
         // Try to get real data first
-        const response = await starService.getAllPlugins?.();
+        const response = await pluginService.getAll();
         return response;
       } catch (error) {
         // Fallback to impressive demo data
